refactor(daemon): extract startChatRun into runtimes/start-chat-run.ts factory
Moves the 3,416-line chat-run closure out of startServer behind an
explicit-deps createStartChatRun(deps) factory. The body is byte-identical
except four daemonUrl references become deps.daemonUrl, passed as a live
getter because server.ts reassigns daemonUrl after listen. 133 captured
identifiers become direct imports; ~39 server.ts module-scope helpers,
registries, and startServer locals flow through deps. server.ts drops from
9,543 to 6,172 lines with an identical export surface.

Verified by the golden SSE characterization suite (4/4), pnpm guard, and
a byte-identity diff against the pre-move closure.
