+++ apps/daemon/src/server.ts
@@ -181,17 +181,11 @@ import {
   getRememberedLiveModels,
   preferFreshLiveModels,
   rememberLiveModels,
-  resolveDefaultModelFromOptions,
   resolveModelForAgent,
 } from './runtimes/models.js';
 import { loadMmdRouteLaunchEnv } from './runtimes/mmd-routes.js';
 import { preparePromptFileForAgent } from './runtimes/prompt-file.js';
-import { TerminalControlSequenceStripper } from './runtimes/terminal-control.js';
 import { buildOpenCodeByokProviderConfig } from './runtimes/byok-opencode.js';
-import {
-  persistPlainStreamArtifacts,
-  plainStdoutFromRunEvents,
-} from './runtimes/plain-stream.js';
 import {
   readVelaLoginStatus,
   resolveAmrProfile,
@@ -603,6 +597,7 @@ import {
 export { rewriteSkillAssetUrls } from './routes/static-resource.js';
 import { registerRoutineRoutes, routineDbRowToContract } from './routes/routine.js';
 import { resolveAmrModelProbe } from './runtimes/amr-model-probe.js';
+import { createStartChatRun } from './runtimes/start-chat-run.js';
 import { createPluginInstallationHelpers, normalizeProjectPluginFolderPath, resolveProjectChildDirectory } from './services/plugin-installation.js';
 import { createPluginShareTaskStore } from './services/plugin-share-tasks.js';
 import { getRouteRegistrationInventory, installRouteRegistrationGuard } from './route-registration-guard.js';
@@ -4009,3557 +4004,51 @@ export async function startServer({
       });
   };
 
-  const startChatRun = async (chatBody, run) => {
-    const lifecycle = createRunLifecycleTracer(run);
-    lifecycle.mark('chat_run_started');
-    /** @type {Partial<ChatRequest> & { imagePaths?: string[] }} */
-    chatBody = chatBody || {};
-    const {
-      agentId,
-      message,
-      currentPrompt,
-      systemPrompt,
-      imagePaths = [],
-      projectId,
-      conversationId,
-      assistantMessageId,
-      clientRequestId,
-      skillId,
-      skillIds,
-      designSystemId,
-      sessionMode,
-      attachments = [],
-      commentAttachments = [],
-      model,
-      reasoning,
-      locale,
-      research,
-      context,
-      titleGeneration,
-      byokProvider,
-      byokMediaDefaults,
-    } = chatBody;
-    lifecycle.mark('prompt_build_start');
-    if (typeof projectId === 'string' && projectId) run.projectId = projectId;
-    if (typeof conversationId === 'string' && conversationId)
-      run.conversationId = conversationId;
-    if (typeof assistantMessageId === 'string' && assistantMessageId)
-      run.assistantMessageId = assistantMessageId;
-    if (typeof clientRequestId === 'string' && clientRequestId)
-      run.clientRequestId = clientRequestId;
-    if (typeof agentId === 'string' && agentId) run.agentId = agentId;
-    // Stash the original user prompt + per-turn config so the
-    // langfuse-bridge report path can include them without reaching back
-    // into chatBody across the createChatRunService boundary. Each field
-    // is optional and only set when the chat body actually carried it.
-    const telemetryPrompt = telemetryPromptFromRunRequest(message, currentPrompt);
-    if (typeof telemetryPrompt === 'string') run.userPrompt = telemetryPrompt;
-    if (typeof model === 'string' && model) run.model = model;
-    if (typeof reasoning === 'string' && reasoning) run.reasoning = reasoning;
-    if (typeof skillId === 'string' && skillId) run.skillId = skillId;
-    if (typeof designSystemId === 'string' && designSystemId)
-      run.designSystemId = designSystemId;
-    const conversationSession =
-      typeof conversationId === 'string' && conversationId
-        ? getConversation(db, conversationId)
-        : null;
-    const runSessionMode =
-      sessionMode === 'chat' || sessionMode === 'design' || sessionMode === 'plan'
-        ? normalizeConversationSessionMode(sessionMode)
-        : normalizeConversationSessionMode(conversationSession?.sessionMode);
-    const def = getAgentDef(agentId);
-    if (!def)
-      return design.runs.fail(
-        run,
-        'AGENT_UNAVAILABLE',
-        `unknown agent: ${agentId}`,
-      );
-    if (!def.bin)
-      return design.runs.fail(run, 'AGENT_UNAVAILABLE', 'agent has no binary');
-    const byokOpenCodeProvider = def.id === 'byok-opencode'
-      ? buildOpenCodeByokProviderConfig(
-          byokProvider,
-          typeof model === 'string' ? model : null,
-        )
-      : null;
-    if (def.id === 'byok-opencode' && !byokOpenCodeProvider) {
-      return design.runs.fail(
-        run,
-        'BYOK_PROVIDER_REQUIRED',
-        'BYOK OpenCode requires a provider, API key, and model for this run.',
-      );
-    }
-    // Validate the checked-in `inactivityTimeoutMs` hint immediately
-    // after the runtime def is selected and before any side-effectful
-    // setup (auto-memory extract, `.mcp.json` write/unlink,
-    // composeSystemPrompt, prompt persistence). A bad def value would
-    // otherwise abort the run only at watchdog-arm time, after that
-    // setup has already mutated local state, leaving confusing partial
-    // residue behind (issue #2467 review on PR #2579).
-    //
-    // Catch is intentionally narrowed to `RangeError`, the only kind
-    // `assertValidRuntimeDefInactivityTimeoutMs` is allowed to throw
-    // for invalid checked-in values. Anything else (a regression that
-    // makes the helper throw on a valid value, an unrelated bug
-    // introduced while touching this path) should bubble up to the
-    // outer chat-run starter — which surfaces it as
-    // `AGENT_EXECUTION_FAILED` — rather than being misreported as
-    // "the runtime def is bad" and burying the real failure.
-    try {
-      assertValidRuntimeDefInactivityTimeoutMs(def.inactivityTimeoutMs);
-    } catch (err) {
-      if (err instanceof RangeError) {
-        return design.runs.fail(run, 'AGENT_RUNTIME_DEF_INVALID', err.message);
-      }
-      throw err;
-    }
-    const safeCommentAttachments =
-      normalizeCommentAttachments(commentAttachments);
-    if (
-      (typeof message !== 'string' || !message.trim()) &&
-      safeCommentAttachments.length === 0
-    ) {
-      return design.runs.fail(run, 'BAD_REQUEST', 'message required');
-    }
-    const browserUseRunState = buildBrowserUseRunState({
-      requested: isBrowserUseRequested(message, currentPrompt, systemPrompt),
-      agentId: def.id,
-    });
-    if (browserUseRunState) {
-      run.browserUse = browserUseRunState;
-      design.runs.emit(run, 'diagnostic', {
-        type: 'browser_use_unavailable',
-        ...browserUseRunState,
-      });
-    }
-    if (run.cancelRequested || design.runs.isTerminal(run.status)) return;
-    const runId = run.id;
-
-    // Auto-memory hook. Pulls explicit "remember:" / "我是 X" / "I prefer Y"
-    // markers out of the just-arrived user message and writes them as MD
-    // files under <dataDir>/memory/. We await so the very next
-    // composeSystemPrompt() call (a few lines below) re-reads memory from
-    // disk and a marker inside this turn's message is reflected in this
-    // turn's prompt. Failures are swallowed — memory is best-effort and
-    // must never block the agent run.
-    if (
-      (run.retryAttemptCount ?? 0) === 0 &&
-      typeof message === 'string' &&
-      message.trim().length > 0
-    ) {
-      try {
-        await extractFromMessage(RUNTIME_DATA_DIR, message);
-      } catch (err) {
-        console.warn('[memory] extractFromMessage failed', err);
-      }
-    }
-
-    // Resolve the project working directory (creating the folder if it
-    // doesn't exist yet). Without one we don't pass cwd to spawn — the
-    // agent then runs in whatever inherited dir, which still lets API
-    // mode work but loses file-tool addressability.
-    // Project directory resolution lives in projects.ts so sandbox mode can
-    // consistently reject imported-folder metadata that has no managed copy.
-    let cwd = null;
-    let existingProjectFiles = [];
-    let existingProjectFolders = [];
-    if (typeof projectId === 'string' && projectId) {
-      try {
-        const chatProject = getProject(db, projectId);
-        const chatMeta = chatProject?.metadata;
-        // ensureProject/resolveProjectDir now resolve external baseDir folders
-        // internally (and assertSandboxProjectRootAvailable rejects imported
-        // folders with no managed copy in sandbox mode), so we pass chatMeta
-        // through instead of branching on baseDir here.
-        assertSandboxProjectRootAvailable(chatMeta);
-        cwd = await ensureProject(PROJECTS_DIR, projectId, chatMeta);
-        existingProjectFiles = await listFiles(PROJECTS_DIR, projectId, { metadata: chatMeta });
-        existingProjectFolders = await listProjectFolders(PROJECTS_DIR, projectId, { metadata: chatMeta });
-      } catch (err) {
-        if (err instanceof SandboxImportedProjectError) {
-          return design.runs.fail(run, 'BAD_REQUEST', err.message);
-        }
-        cwd = null;
-        existingProjectFiles = [];
-        existingProjectFolders = [];
-      }
-    }
-    if (run.cancelRequested || design.runs.isTerminal(run.status)) return;
-
-    // Sanitise supplied image paths: must live under UPLOAD_DIR and stay
-    // below the prompt-image safety cap.
-    const { safeImages, oversizedImages, failedImages } =
-      resolveSafePromptImagePaths(imagePaths);
-    if (oversizedImages.length > 0) {
-      return design.runs.fail(
-        run,
-        'BAD_REQUEST',
-        'Image attachments must be 1 MB or smaller.',
-      );
-    }
-    if (failedImages.length > 0) {
-      return design.runs.fail(
-        run,
-        'INTERNAL_ERROR',
-        'Failed to read one or more image attachments.',
-      );
-    }
-    const amrStagedImages =
-      def.id === 'amr'
-        ? await stageAmrImagePaths(cwd ?? PROJECT_ROOT, safeImages, UPLOAD_DIR)
-        : safeImages;
-
-    // Project-scoped attachments: project-relative paths inside cwd. Each
-    // is run through the same path-traversal guard the file CRUD endpoints
-    // use, then existence-checked. Whatever survives shows up as an
-    // explicit list at the bottom of the user message so the agent knows
-    // to Read it.
-    const safeAttachments = cwd
-      ? resolveSafeProjectAttachments(cwd, attachments)
-      : [];
-    run.projectAttachmentPaths = safeAttachments;
-
-    // Local code agents don't accept a separate "system" channel the way the
-    // Messages API does — we fold the skill + design-system prompt into the
-    // user message. The <artifact> wrapping instruction comes from
-    // systemPrompt. We also stitch in the cwd hint so the agent knows
-    // where its file tools should write, and the attachment list so it
-    // doesn't have to guess what the user just dropped in.
-    const projectRecord =
-      typeof projectId === 'string' && projectId
-        ? getProject(db, projectId)
-        : null;
-    const runContextPrompt = renderRunContextPrompt(context, projectRecord?.metadata);
-    const linkedDirs = (() => {
-      if (!Array.isArray(projectRecord?.metadata?.linkedDirs)) return [];
-      const v = validateLinkedDirs(projectRecord.metadata.linkedDirs);
-      return v.dirs ?? [];
-    })();
-    const cwdHint = cwd
-      ? formatDesignFilesWorkspaceHint(cwd, existingProjectFiles, existingProjectFolders)
-      : '';
-    const linkedDirsHint = linkedDirs.length > 0
-      ? `\n\nLinked code folders (read-only reference code the user wants you to see):\n${
-          linkedDirs.map((d) => `- \`${d}\``).join('\n')
-        }`
-      : '';
-    const attachmentHint = formatProjectAttachmentHint(safeAttachments);
-    // Plan §3.A3 / spec §9: thread plugin context onto every tool token
-    // so the connector execute route can re-validate the §5.3
-    // capability gate without re-reading the SQLite snapshot row.
-    let pluginGrantContext = null;
-    if (cwd && typeof projectId === 'string' && projectId && run?.appliedPluginSnapshotId) {
-      const snap = getSnapshot(db, run.appliedPluginSnapshotId);
-      if (snap) {
-        const installed = getInstalledPlugin(db, snap.pluginId);
-        pluginGrantContext = {
-          pluginSnapshotId: snap.snapshotId,
-          pluginTrust: installed?.trust ?? 'restricted',
-          pluginCapabilitiesGranted: snap.capabilitiesGranted ?? [],
-        };
-      }
-    }
-    const toolTokenGrant = cwd && typeof projectId === 'string' && projectId
-      ? toolTokenRegistry.mint({
-          runId,
-          projectId,
-          allowedEndpoints: CHAT_TOOL_ENDPOINTS,
-          allowedOperations: CHAT_TOOL_OPERATIONS,
-          ...(pluginGrantContext ?? {}),
-        })
-      : null;
-    let toolTokenRevoked = false;
-    const revokeToolToken = (reason) => {
-      if (toolTokenRevoked || !toolTokenGrant) return;
-      toolTokenRevoked = true;
-      toolTokenRegistry.revokeToken(toolTokenGrant.token, reason);
-    };
-    // The async startup phase below (compose prompt, prepare prompt file,
-    // probe models, …) has many awaits and no blanket try/finally; an
-    // exception there finalizes the run via runs.fail() without running the
-    // per-attempt cleanup wired to the child lifecycle. Register the grant +
-    // sink release on the run's terminal chokepoint so any exit path — startup
-    // throw included — revokes the capability token instead of leaking it for
-    // the token TTL. Idempotent with the explicit pre-spawn/child-close cleanup.
-    if (toolTokenGrant) {
-      run.onFinalize = () => {
-        revokeToolToken('run_finalized');
-        const sinkRunId = toolTokenGrant.runId ?? runId;
-        activeChatAgentEventSinks.delete(sinkRunId);
-        activeChatRunHandles.delete(sinkRunId);
-      };
-    }
-    const runtimeToolPrompt = createAgentRuntimeToolPrompt(daemonUrl, toolTokenGrant);
-    const commentHint = renderCommentAttachmentHint(safeCommentAttachments);
-
-    // Resolve external MCP config + stored OAuth tokens up-front so the
-    // system prompt can warn the model away from Claude Code's synthetic
-    // `*_authenticate` / `*_complete_authentication` tools for any
-    // server the daemon already holds a valid Bearer for. We re-use both
-    // values further down at .mcp.json write time — see the spawn block
-    // below — instead of re-reading.
-    let externalMcpConfig = { servers: [] };
-    if (!SANDBOX_RUNTIME.enabled) {
-      try {
-        externalMcpConfig = await readMcpConfig(RUNTIME_DATA_DIR);
-      } catch (err) {
-        console.warn(
-          '[mcp-config] read failed:',
-          err && err.message ? err.message : err,
-        );
-      }
-    }
-    const runScopedMcpServers = Array.isArray(run?.toolBundle?.mcpServers)
-      ? run.toolBundle.mcpServers
-      : [];
-    const {
-      enabledServers: enabledExternalMcp,
-      persistedTokenServerIds,
-    } = resolveExternalMcpServersForRun({
-      persistedServers: externalMcpConfig.servers,
-      runScopedServers: runScopedMcpServers,
-      sandboxMode: SANDBOX_RUNTIME.enabled,
-    });
-    const oauthTokensForSpawn = {};
-    if (persistedTokenServerIds.size > 0) {
-      try {
-        const stored = await readAllTokens(RUNTIME_DATA_DIR);
-        for (const [serverId, tok] of Object.entries(stored)) {
-          if (!persistedTokenServerIds.has(serverId)) continue;
-          // Default to the persisted access token; null it out if expired so
-          // we never inject a stale `Authorization: Bearer …` header. The
-          // model treats a server with a Bearer pinned as connected and
-          // discourages re-auth, which is the worst possible UX when the
-          // token is going to 401 every call.
-          let access = isTokenExpired(tok) ? null : tok.accessToken;
-          if (isTokenExpired(tok) && tok.refreshToken) {
-            try {
-              const refreshed = await refreshAndPersistToken(
-                RUNTIME_DATA_DIR,
-                serverId,
-                tok,
-              );
-              if (refreshed) access = refreshed.accessToken;
-            } catch (err) {
-              console.warn(
-                '[mcp-oauth] refresh failed for',
-                serverId,
-                err && err.message ? err.message : err,
-              );
-            }
-          }
-          if (access) {
-            oauthTokensForSpawn[serverId] = access;
-          } else {
-            console.warn(
-              '[mcp-oauth] skipping expired token for',
-              serverId,
-              '— reconnect required',
-            );
-          }
-        }
-      } catch (err) {
-        console.warn(
-          '[mcp-tokens] read failed:',
-          err && err.message ? err.message : err,
-        );
-      }
-    }
-    const connectedExternalMcp = enabledExternalMcp
-      .filter((s) => typeof oauthTokensForSpawn[s.id] === 'string')
-      .map((s) => ({ id: s.id, label: s.label }));
-
-    const {
-      prompt: daemonSystemPrompt,
-      activeSkillDirs,
-      critiqueShouldRun,
-      designSystemSelection,
-      promptTelemetryParts,
-    } =
-      await composeDaemonSystemPrompt({
-        agentId,
-        projectId,
-        skillId,
-        skillIds,
-        designSystemId,
-        streamFormat: def?.streamFormat ?? 'plain',
-        locale,
-        sessionMode: runSessionMode,
-        connectedExternalMcp,
-        mediaExecution: run?.mediaExecution,
-        byokMediaDefaults,
-        // Plan §3.M2 / §3.V1 — forward the run's snapshot id so the
-        // prompt composer can splice in `## Active stage` blocks.
-        // Default ON; set OD_BUNDLED_ATOM_PROMPTS=0 to opt out.
-        appliedPluginSnapshotId: run?.appliedPluginSnapshotId ?? null,
-      });
-
-    run.designSystemId = designSystemSelection?.id ?? null;
-    run.designSystemRequestedId = designSystemSelection?.requestedId ?? null;
-    run.designSystemSelectionSource = designSystemSelection?.source ?? 'none';
-    run.designSystemDigest = designSystemSelection?.digest ?? null;
-
-    // Make skill side files reachable through three layers, in order of
-    // preference. The skill preamble emitted by `withSkillRootPreamble()`
-    // advertises both the cwd-relative path (1) and the absolute path
-    // (2/3) so the agent can pick whichever works.
-    //
-    //   1. CWD-relative copy. Stage every active/composed skill into
-    //      `<cwd>/.od-skills/<folder>/` so any agent CLI — not just the
-    //      ones that honour `--add-dir` — can reach those files via a
-    //      path inside its working directory. We copy (not symlink) so
-    //      each staged directory is a true write barrier — agents cannot
-    //      mutate the shipped repo resource through their cwd.
-    //   2. `--add-dir` allowlist. For non-Codex agents, pass `SKILLS_DIR`
-    //      and `DESIGN_SYSTEMS_DIR` so the absolute fallback path in the
-    //      preamble is reachable when staging fails (e.g. the project has
-    //      no on-disk cwd, or fs.cp errored). Codex treats `--add-dir`
-    //      entries as writable, so Codex receives only the narrow
-    //      `${CODEX_HOME:-$HOME/.codex}/generated_images` output folder
-    //      for allowlisted gpt-image image projects.
-    //   3. PROJECT_ROOT cwd. When `cwd` is null, the agent runs with
-    //      `cwd: PROJECT_ROOT` — there the absolute path is already an
-    //      in-cwd path, so neither (1) nor (2) is required for it to
-    //      resolve.
-    //
-    // Design systems are *not* staged here. Their bodies are read by the
-    // daemon and folded into the system prompt directly (see
-    // `readDesignSystem`), so an agent never has to open them via the
-    // filesystem.
-    if (cwd && activeSkillDirs.length > 0) {
-      for (const skillDir of activeSkillDirs) {
-        const result = await stageActiveSkill(
-          cwd,
-          skillCwdAliasSegment(skillDir),
-          skillDir,
-          (msg) => console.warn(msg),
-        );
-        if (!result.staged) {
-          console.warn(
-            `[od] skill-stage skipped: ${result.reason ?? 'unknown reason'}; falling back to absolute paths`,
-          );
-        }
-      }
-    }
-    // Resolve the agent's effective working directory once and use it
-    // everywhere the agent could read it (buildArgs runtimeContext, spawn
-    // cwd, ACP session new). Falling back to PROJECT_ROOT — rather than
-    // letting `spawn` inherit the daemon process cwd — is what makes the
-    // absolute-path fallback in the skill preamble actually in-cwd for
-    // no-project runs (packaged daemons / service launches do not start
-    // their working directory from the workspace root).
-    const effectiveCwd = cwd ?? PROJECT_ROOT;
-    // Baseline the project's artifact files before the agent runs, so the
-    // run-finished handler can diff against them and report `artifact_count`
-    // for ANY agent (not just claude_code). Only for real project runs: a
-    // null `cwd` means a no-project run rooted at PROJECT_ROOT, whose churn is
-    // not the user's artifacts — those fall back to the tool-stream count.
-    if (run?.id && cwd) {
-      try {
-        runArtifactBaselines.remember(run.id, cwd, snapshotProjectArtifacts(cwd));
-      } catch {
-        // Snapshotting is best-effort; finish falls back to the tool-stream count.
-      }
-    }
-    const latestRunPromptForHtmlVersionSnapshot = () => {
-      if (run.conversationId) {
-        try {
-          const row = db.prepare(
-            `SELECT content
-               FROM messages
-              WHERE conversation_id = ?
-                AND role = 'user'
-                AND LENGTH(TRIM(content)) > 0
-              ORDER BY COALESCE(ended_at, started_at, created_at, 0) DESC,
-                       position DESC
-              LIMIT 1`,
-          ).get(run.conversationId);
-          if (typeof row?.content === 'string' && row.content.trim()) {
-            return { prompt: row.content.trim(), promptSource: 'message' as const };
-          }
-        } catch {
-          // Version prompt provenance is best-effort.
-        }
-      }
-      const requestPrompt =
-        typeof currentPrompt === 'string' && currentPrompt.trim()
-          ? currentPrompt.trim()
-          : typeof message === 'string' && message.trim()
-            ? message.trim()
-            : null;
-      return requestPrompt ? { prompt: requestPrompt, promptSource: 'message' as const } : { prompt: null };
-    };
-    const snapshotAiHtmlVersionsBeforeSuccess = async () => {
-      if (!run?.id || !run.projectId) return;
-      const artifactBaseline = runArtifactBaselines.peek(run.id);
-      if (!artifactBaseline || artifactBaseline.contended) return;
-      let diff;
-      try {
-        diff = diffRunArtifacts(
-          artifactBaseline.before,
-          snapshotProjectArtifacts(artifactBaseline.cwd),
-        );
-      } catch {
-        return;
-      }
-      const promptInfo = latestRunPromptForHtmlVersionSnapshot();
-      await snapshotAiHtmlVersionsForRun({
-        projectsRoot: PROJECTS_DIR,
-        projectId: run.projectId,
-        projectRoot: artifactBaseline.cwd,
-        diff,
-        prompt: promptInfo.prompt,
-        ...(promptInfo.promptSource ? { promptSource: promptInfo.promptSource } : {}),
-        metadata: projectRecord?.metadata,
-      });
-    };
-    let codexGeneratedImagesDir = resolveCodexGeneratedImagesDir(
-      agentId,
-      projectRecord?.metadata,
-      process.env,
-      os.homedir(),
-      run?.mediaExecution,
-    );
-    if (codexGeneratedImagesDir) {
-      codexGeneratedImagesDir = validateCodexGeneratedImagesDir(
-        codexGeneratedImagesDir,
-        {
-          protectedDirs: [SKILLS_DIR, DESIGN_SYSTEMS_DIR, ...linkedDirs],
-        },
-      );
-    }
-    const extraAllowedDirs = resolveChatExtraAllowedDirs({
-      agentId,
-      skillsDir: SKILLS_DIR,
-      designSystemsDir: DESIGN_SYSTEMS_DIR,
-      linkedDirs,
-      codexGeneratedImagesDir,
-    });
-    const codexImagegenOverride = resolveGrantedCodexImagegenOverride({
-      agentId,
-      metadata: projectRecord?.metadata,
-      codexGeneratedImagesDir,
-      extraAllowedDirs,
-      mediaExecution: run?.mediaExecution,
-    });
-    const researchCommandContract = resolveResearchCommandContract(
-      research,
-      message,
-    );
-    // Resume-capable adapters continue their own upstream session so they
-    // keep working memory across turns. Decide once per run; reuse for the
-    // prompt-composition skipTranscript choice, the buildArgs flags, and the
-    // create-turn persistence below.
-    const agentSupportsSessionResume =
-      def.resumesSessionViaCli === true ||
-      def.streamFormat === 'pi-rpc' ||
-      def.resumesSessionViaAcpLoad === true;
-    // Capture-style adapters (codex) mint their OWN session id and report it on
-    // the stream; the daemon captures it here and persists THAT as the resume
-    // handle instead of `agentResumeCtx.newSessionId` (which such CLIs ignore).
-    // Set from the `status` event's `sessionId` in `sendAgentEvent` below.
-    const agentCapturesSessionId = def.capturesSessionIdFromStream === true;
-    let capturedSessionId: string | null = null;
-    // --- Model resolution hoisted above the resume-identity guard ---
-    // The guard (and the persisted `agent_sessions.model`) must key off the
-    // model identity actually requested for this turn. Explicit `default` is
-    // kept as a real identity because ACP runtimes can leave model selection to
-    // the upstream session's own configured default; omitted models may still
-    // resolve to an available fallback below.
-    let configuredAgentEnv = {};
-    try {
-      const appConfig = await readAppConfig(RUNTIME_DATA_DIR);
-      configuredAgentEnv = agentCliEnvForAgent(appConfig.agentCliEnv, def.id);
-    } catch {
-      configuredAgentEnv = {};
-    }
-    const requestedLiveModelScope = def.id === 'amr'
-      ? resolveAmrProfile({
-          ...process.env,
-          ...(def.env || {}),
-          ...configuredAgentEnv,
-        })
-      : null;
-    let safeModel = resolveModelForAgent(
-      def,
-      typeof model === 'string'
-        ? isKnownModel(def, model, requestedLiveModelScope)
-          ? model
-          : sanitizeCustomModel(model)
-        : null,
-      process.env,
-      requestedLiveModelScope,
-    );
-    const hasDefaultModelEnvOverride = Boolean(
-      def.defaultModelEnvVar &&
-      typeof process.env[def.defaultModelEnvVar] === 'string' &&
-      process.env[def.defaultModelEnvVar]?.trim(),
-    );
-    const safeReasoning =
-      typeof reasoning === 'string' && Array.isArray(def.reasoningOptions)
-        ? (def.reasoningOptions.find((r) => r.id === reasoning)?.id ?? null)
-        : null;
-    const agentOptions = { model: safeModel, reasoning: safeReasoning };
-    const agentLaunch = resolveAgentLaunch(def, configuredAgentEnv);
-    const resolvedBin = agentLaunch.selectedPath;
-    if (def.id === 'amr' && resolvedBin && agentLaunch.launchPath) {
-      // Concretize omitted/default AMR model requests to the live catalog
-      // default before the resume guard. The AMR preflight below applies the
-      // same rewrite before spawn; keeping this earlier copy aligned prevents
-      // stored concrete session models from comparing against raw `default`.
-      try {
-        const resumeProbe = await resolveAmrModelProbe({ dataDir: RUNTIME_DATA_DIR, env: process.env, readAppConfig });
-        const resumeCatalog = await amrModelLoadingCache.get(resumeProbe.cacheKey, {
-          fetchPreset: () => fetchVelaPresetModels(resumeProbe.launchPath, resumeProbe.env),
-          fetchRemote: () => fetchVelaRemoteModelsWithRetry(resumeProbe.launchPath, resumeProbe.env),
-        });
-        const resumeLiveModels = preferFreshLiveModels(
-          resumeCatalog.models ?? [],
-          getRememberedLiveModels(def.id, requestedLiveModelScope),
-        );
-        const resumeModelIds = new Set(resumeLiveModels.map((c) => c?.id).filter(Boolean));
-        const askedForDefault =
-          typeof model !== 'string' || !model.trim() || model.trim().toLowerCase() === 'default';
-        const defaultRunModel = resolveDefaultModelFromOptions(resumeLiveModels);
-        if (
-          !safeModel ||
-          safeModel === 'default' ||
-          (
-            askedForDefault &&
-            !hasDefaultModelEnvOverride &&
-            defaultRunModel &&
-            (!resumeModelIds.has(safeModel) || safeModel !== defaultRunModel)
-          )
-        ) {
-          safeModel = defaultRunModel ?? safeModel ?? null;
-          agentOptions.model = safeModel;
-        }
-      } catch {
-        // Degrade silently: keep the requested value. The preflight below records
-        // the probe failure and applies the identical fallback.
-      }
-    }
-    const agentResumeCtx =
-      agentSupportsSessionResume && run.conversationId
-        ? resolveAgentResumeContext(db, {
-            conversationId: run.conversationId,
-            agentId: def.id,
-            currentModel: safeModel ?? null,
-            currentCwd: effectiveCwd,
-            currentAssistantMessageId: run.assistantMessageId ?? null,
-          })
-        : { resumeSessionId: null as string | null, newSessionId: undefined as string | undefined, isResuming: false, storedStablePromptHash: null as string | null, invalidationReason: null };
-    const userRequestPrompt = composeChatUserRequestForAgent(
-      message,
-      currentPrompt,
-      // Only trim to the latest turn when we are actually resuming an
-      // existing session. A create turn still sends the full transcript so
-      // a brand-new session (incl. first turn after another agent)
-      // is seeded with prior context.
-      { skipTranscript: agentResumeCtx.isResuming },
-    );
-    // The stable instruction slice (daemon prompt + tool contract + system
-    // prompt = design system / skills / memory) is identical across turns of
-    // a conversation in the common case. A resumed Claude session already
-    // holds it, so on resume turns we skip it unless it changed since the
-    // session was seeded — keyed by a hash stored on agent_sessions. Create
-    // turns and changed-hash turns send the full block (byte-identical to the
-    // previous behavior); non-resume agents have isResuming === false and so
-    // always send the full block.
-    const stableInstructionFingerprint = [daemonSystemPrompt, runtimeToolPrompt, systemPrompt]
-      .map((part) => (typeof part === 'string' ? part.trim() : ''))
-      .join('\n\n---\n\n');
-    const currentStableHash = hashStableInstructions(stableInstructionFingerprint);
-    // `runtimeToolPrompt` is part of the fingerprint and varies only when the
-    // tool-token grant's presence flips between turns (rare cwd/projectId edge
-    // cases); any such change correctly forces a full re-send that turn.
-    const includeStableInstructions = computeIncludeStable(
-      agentResumeCtx.isResuming,
-      agentResumeCtx.storedStablePromptHash,
-      currentStableHash,
-    );
-    run.promptCache = describeStablePromptCache({
-      isResuming: agentResumeCtx.isResuming,
-      storedStablePromptHash: agentResumeCtx.storedStablePromptHash,
-      currentStableHash,
-    });
-    const browserUsePromptGuard = renderBrowserUseUnavailablePrompt(run.browserUse ?? null);
-    const titleGenerationRequested =
-      titleGeneration &&
-      typeof titleGeneration === 'object' &&
-      titleGeneration.enabled === true &&
-      !agentResumeCtx.isResuming;
-    const titleGenerationPrompt = titleGenerationRequested
-      ? [
-          'Internal title task:',
-          'Before answering the user request, emit exactly one short title marker:',
-          '<od-title>Title Here</od-title>',
-          'Rules: 2-6 words, preserve the user request language, no quotes, no markdown, no punctuation unless necessary.',
-          'Do not mention this title task to the user. Continue with the normal answer after the title marker.',
-        ].join('\n')
-      : '';
-    const clientInstructionParts = includeStableInstructions
-      ? [researchCommandContract, runContextPrompt, browserUsePromptGuard, titleGenerationPrompt, systemPrompt]
-      : [researchCommandContract, runContextPrompt, browserUsePromptGuard, titleGenerationPrompt];
-    const clientInstructionPrompt = clientInstructionParts
-      .map((part) => (typeof part === 'string' ? part.trim() : ''))
-      .filter(Boolean)
-      .join('\n\n---\n\n');
-    const instructionPrompt = composeLiveInstructionPrompt({
-      daemonSystemPrompt: includeStableInstructions ? daemonSystemPrompt : '',
-      runtimeToolPrompt: includeStableInstructions ? runtimeToolPrompt : '',
-      clientSystemPrompt: clientInstructionPrompt,
-      finalPromptOverride: codexImagegenOverride,
-    });
-    // Some models (notably claude-opus-4-7 with --include-partial-messages)
-    // start their reply by echoing the top of the user message verbatim,
-    // so the rendered chat shows a "# Instructions ..." block ahead of the
-    // real answer. Closing every Instructions block with an explicit
-    // "do not echo" line cuts the regression in practice without changing
-    // the turn-shape every agent CLI expects (user message carrying both
-    // instructions and request) — see server.ts:9920 composer notes.
-    const ECHO_GUARD =
-      '\n\n(Do not quote, restate, or echo the # Instructions block above in your reply. Begin your response with the answer to the # User request below.)';
-    const formAnswerMatch = FORM_ANSWERS_HEADER_RE.exec(
-      typeof currentPrompt === 'string' ? currentPrompt : '',
-    );
-    const formIdForOverride = formAnswerMatch
-      ? ((formAnswerMatch[1] || 'form').trim().replace(/[^\w.-]/g, '') || 'form').toLowerCase()
-      : null;
-    const formOverride =
-      formIdForOverride === 'discovery' || formIdForOverride === 'task-type'
-        ? FORM_ANSWERED_SYSTEM_OVERRIDE
-        : formIdForOverride !== null
-          ? FORM_ANSWERED_GENERIC_OVERRIDE
-          : '';
-    const promptImagePaths = selectPromptImagePaths(
-      def.id,
-      safeImages,
-      amrStagedImages,
-    );
-    const composed = [
-      instructionPrompt
-        ? `# Instructions (read first)\n\n${formOverride}${instructionPrompt}${cwdHint}${linkedDirsHint}${ECHO_GUARD}\n\n---\n`
-        : cwdHint
-          ? `# Instructions\n\n${formOverride}${cwdHint}${linkedDirsHint}${ECHO_GUARD}\n\n---\n`
-          : linkedDirsHint
-            ? `# Instructions\n\n${formOverride}${linkedDirsHint}${ECHO_GUARD}\n\n---\n`
-            : formOverride
-              ? `# Instructions\n\n${formOverride}${ECHO_GUARD}\n\n---\n`
-              : '',
-      `# User request\n\n${userRequestPrompt}${attachmentHint}${commentHint}`,
-      promptImagePaths.length
-        ? `\n\n${promptImagePaths.map((p) => `@${p}`).join(' ')}`
-        : '',
-    ].join('');
-    run.promptTelemetry = buildPromptStackTelemetry({
-      composedPrompt: composed,
-      sections: [
-        { kind: 'formOverride', content: formOverride },
-        // Phase 1 explicitly needs redactedContent for these aggregate prompts:
-        // they are the quickest way to inspect the system context sent to the
-        // model when diagnosing Langfuse traces.
-        { kind: 'daemonSystemPrompt', content: daemonSystemPrompt },
-        { kind: 'runtimeToolPrompt', content: runtimeToolPrompt },
-        { kind: 'researchCommandContract', content: researchCommandContract },
-        { kind: 'runContextPrompt', content: runContextPrompt },
-        { kind: 'browserUsePromptGuard', content: browserUsePromptGuard },
-        { kind: 'clientSystemPrompt', content: clientInstructionPrompt },
-        { kind: 'echoGuard', content: ECHO_GUARD },
-        { kind: 'userRequest', content: userRequestPrompt },
-        { kind: 'skillPrompt', content: promptTelemetryParts?.skillPrompt },
-        {
-          kind: 'designSystemPrompt',
-          content: promptTelemetryParts?.designSystemPrompt,
-        },
-        {
-          kind: 'pluginStagePrompt',
-          content: promptTelemetryParts?.pluginStagePrompt,
-        },
-        { kind: 'cwdHint', content: cwdHint, metadata: cwd ? [cwd] : [] },
-        {
-          kind: 'linkedDirsHint',
-          content: linkedDirsHint,
-          metadata: linkedDirs,
-        },
-        {
-          kind: 'attachments',
-          content: attachmentHint,
-          metadata: safeAttachments,
-        },
-        {
-          kind: 'commentAttachments',
-          content: commentHint,
-          metadata: safeCommentAttachments,
-        },
-        {
-          kind: 'promptImagePaths',
-          content: promptImagePaths.join('\n'),
-          metadata: promptImagePaths,
-        },
-      ],
-    });
-    lifecycle.mark('prompt_build_end');
-    lifecycle.mark('launch_preflight_start');
-    // (model resolution + AMR concretization hoisted above the resume guard)
-    const executionProfile = executionProfileFromStreamFormat(def.streamFormat);
-    // Accumulates the agent's visible text this run so the close handler can
-    // tell whether the turn ended on a clarifying question form. The
-    // `od-plugin-authoring` plugin's turn-1 flow is to emit a
-    // `<question-form>` collecting the plugin brief, then STOP and wait for
-    // the user to answer (see the `discovery-question-form` atom in
-    // `plugins/scaffold.ts`). That turn legitimately closes with `code === 0`
-    // and no `generated-plugin/` artifacts yet, so the missing-artifacts
-    // guard must not treat it as a failure. We buffer the streamed text
-    // rather than read the persisted message because the assistant message
-    // row may not be wired up at close time. The buffer is capped because a
-    // discovery form streams near the top of the turn; we only need enough to
-    // validate the first complete form block (see
-    // `emittedRenderableQuestionForm`).
-    const CLARIFYING_QUESTION_BUFFER_CAP = 256 * 1024;
-    let clarifyingQuestionText = '';
-    let visibleAssistantText = '';
-    // Reply text handed to the background memory extractor at child-close.
-    // Captures the GUARDED, visible reply from BOTH channels a run can emit on:
-    // structured agents' `agent` `text_delta` (Claude/Codex/Gemini/Copilot/ACP/
-    // qoder/pi-rpc) and the plain/BYOK/antigravity family's `stdout` chunks. So
-    // every agent family contributes its actual reply, and none leak raw
-    // transport frames (system:init, stream_event, hooks). Kept separate from
-    // `visibleAssistantText` so the filesystem empty-output guard that reads
-    // that variable keeps its text_delta-only semantics. Bounded — the
-    // extractor only needs the head of the reply.
-    const MEMORY_REPLY_CAP = 32 * 1024;
-    let memoryReplyText = '';
-    const send = (event, data) => {
-      const lifecycleMarkers = runLifecycleMarkersForStreamEvent(event, data);
-      if (lifecycleMarkers.firstModelEventType) {
-        lifecycle.markFirstModelEvent(lifecycleMarkers.firstModelEventType);
-      }
-      if (lifecycleMarkers.firstVisibleOutput) {
-        lifecycle.mark('first_visible_output');
-      }
-      if (lifecycleMarkers.firstArtifactWrite) {
-        lifecycle.mark('first_artifact_write');
-      }
-      if (
-        event === 'agent' &&
-        data &&
-        data.type === 'text_delta' &&
-        typeof data.delta === 'string' &&
-        clarifyingQuestionText.length < CLARIFYING_QUESTION_BUFFER_CAP
-      ) {
-        clarifyingQuestionText = (clarifyingQuestionText + data.delta).slice(
-          0,
-          CLARIFYING_QUESTION_BUFFER_CAP,
-        );
-      }
-      if (
-        event === 'agent' &&
-        data &&
-        data.type === 'text_delta' &&
-        typeof data.delta === 'string'
-      ) {
-        visibleAssistantText += data.delta;
-      }
-      // Accumulate the visible reply for the memory extractor from whichever
-      // channel this agent family uses: `agent` text_delta (structured streams)
-      // or `stdout` chunks (plain/BYOK/antigravity). Both carry already-guarded,
-      // user-visible text, so this never captures thinking, tool traffic, or raw
-      // transport frames.
-      if (memoryReplyText.length < MEMORY_REPLY_CAP) {
-        const replyPiece =
-          event === 'agent' && data && data.type === 'text_delta' && typeof data.delta === 'string'
-            ? data.delta
-            : event === 'stdout' && data && typeof data.chunk === 'string'
-              ? data.chunk
-              : '';
-        if (replyPiece) {
-          memoryReplyText = (memoryReplyText + replyPiece).slice(0, MEMORY_REPLY_CAP);
-        }
-      }
-      persistRunEventToAssistantMessage(db, run, event, data);
-      design.runs.emit(run, event, data);
-    };
-    const retryAnalyticsBase = (decision, failure, errorCode) => {
-      const runProjectKind = resolveRunProjectKindForAnalytics({
-        hintProjectKind: null,
-        projectMetadata: projectRecord?.metadata,
-      });
-      const isDesignSystemRun =
-        runProjectKind === 'design_system' ||
-        (typeof designSystemId === 'string' && designSystemId.length > 0);
-      return {
-        page_name: isDesignSystemRun ? 'design_system_project' : 'chat_panel',
-        area: isDesignSystemRun ? 'design_system_generation' : 'chat_panel',
-        project_id: typeof projectId === 'string' ? projectId : run.projectId,
-        conversation_id:
-          typeof conversationId === 'string' ? conversationId : run.conversationId ?? null,
-        run_id: run.id,
-        retry_of_run_id: run.id,
-        retry_attempt_index: decision.retryAttemptIndex,
-        retry_max_attempts: decision.retryMaxAttempts,
-        retry_strategy: decision.retryStrategy,
-        agent_provider_id: agentIdToTracking(agentId),
-        model_id: modelIdForTracking(safeModel ?? model),
-        ...(failure?.failure_category ? { failure_category: failure.failure_category } : {}),
-        ...(failure?.failure_detail ? { failure_detail: failure.failure_detail } : {}),
-        ...(failure?.failure_stage ? { failure_stage: failure.failure_stage } : {}),
-        ...(errorCode ? { error_code: errorCode } : {}),
-      };
-    };
-    const destroyChildStdio = (child) => {
-      // Best-effort cleanup of stdio streams on a child process we're about
-      // to drop. The daemon-sidecar (apps/daemon) keeps listeners attached
-      // to child.stdout / child.stderr / child.stdin across the run
-      // lifecycle (line ~12890..~13500+ in this file). Those listeners hold
-      // the Stream objects alive, and the Stream objects own the read side
-      // of the OS pipes — so dropping the child reference via
-      // `run.child = null` without destroying the streams leaks the pipe
-      // file descriptors. After a few hundred retries the daemon
-      // accumulates 10k+ FDs and posix_spawn returns EBADF.
-      //
-      // See: https://github.com/nexu-io/open-design/issues/4100
-      if (!child) return;
-      const destroyStream = (stream) => {
-        if (!stream || stream.destroyed) return;
-        try { stream.removeAllListeners(); } catch {}
-        try { stream.destroy(); } catch {}
-      };
-      destroyStream(child.stdout);
-      destroyStream(child.stderr);
-      destroyStream(child.stdin);
-    };
-    // Synchronously detach the failed attempt: kill the old child and move the
-    // run back to `queued` *now*, even when the re-spawn is delayed by backoff.
-    // This must not be deferred — leaving the old child alive during the backoff
-    // window lets a follow-on signal (e.g. the inactivity watchdog's SIGTERM)
-    // drive a second close-handler pass that finalizes the run as failed before
-    // the retry ever spawns.
-    const tearDownAttemptForRetry = () => {
-      // Release the previous child's stdio streams before letting the
-      // reference drop — see destroyChildStdio for rationale.
-      destroyChildStdio(run.child);
-      if (
-        run.child &&
-        typeof run.child.kill === 'function' &&
-        run.child.exitCode === null &&
-        !run.child.killed
-      ) {
-        try { run.child.kill('SIGTERM'); } catch {}
-      }
-      run.status = 'queued';
-      run.updatedAt = Date.now();
-      run.child = null;
-      run.acpSession = null;
-      run.exitCode = null;
-      run.signal = null;
-      run.error = null;
-      run.errorCode = null;
-      run.stdinOpen = false;
-      lifecycle.resetForAttempt(run.retryAttemptCount ?? 0);
-      run.analyticsTelemetry = {
-        startRequestedAt: run.analyticsTelemetry?.startRequestedAt ?? run.createdAt,
-      };
-    };
-    const spawnRetryAttempt = () => {
-      void startChatRun(chatBody, run).catch((err) => {
-        const message = err instanceof Error ? err.message : String(err);
-        design.runs.emit(
-          run,
-          'error',
-          createSseErrorPayload('AGENT_EXECUTION_FAILED', message),
-        );
-        // Route the retried-start failure through the same finalizer as child
-        // close/error so it emits terminal retry telemetry (run_retry_finished
-        // with retry_result: 'failed') and sets run.retryFinalResult, instead
-        // of finishing directly and leaving run_finished to report the fallback
-        // retry_final_result: 'not_attempted'. retryAttemptCount is already 1
-        // here, so decideSafeRunRetry suppresses with attempt_limit_reached and
-        // cannot trigger another restart loop.
-        finishWithRetryDecision('failed', 1, null);
-      });
-    };
-    // Tear the failed attempt down now (moving the run to `queued`), then wait
-    // out the policy's backoff before re-spawning. Stays cancel-aware: a cancel
-    // or shutdown during the backoff window clears the timer (runtimes/runs.ts)
-    // and finalizes the queued run, and the callback re-checks cancel/terminal
-    // state in case it fires first.
-    const scheduleRetryRestart = (delayMs) => {
-      tearDownAttemptForRetry();
-      const wait = Number.isFinite(delayMs) && delayMs > 0 ? delayMs : 0;
-      if (wait <= 0) {
-        spawnRetryAttempt();
-        return;
-      }
-      run.retryRestartTimer = setTimeout(() => {
-        run.retryRestartTimer = null;
-        if (run.cancelRequested || design.runs.isTerminal(run.status)) return;
-        spawnRetryAttempt();
-      }, wait);
-    };
-    const finalizeRetryTelemetry = (status, decision, failure, errorCode) => {
-      const attemptCount = run.retryAttemptCount ?? 0;
-      const result = runResultFromStatus(status);
-      if (attemptCount <= 0 && result !== 'failed') {
-        run.retryFinalResult = 'not_attempted';
-        run.retrySuppressedReason = undefined;
-        return;
-      }
-      const retryResult =
-        attemptCount > 0
-          ? result === 'success'
-            ? 'success'
-            : result === 'failed'
-              ? 'failed'
-              : 'suppressed'
-          : 'suppressed';
-      const retrySuppressedReason =
-        retryResult === 'suppressed'
-          ? run.cancelRequested
-            ? 'cancel_requested'
-            : decision?.retrySuppressedReason
-          : undefined;
-      const eventDecision =
-        attemptCount > 0
-          ? { ...decision, retryAttemptIndex: attemptCount }
-          : decision;
-      run.retryFinalResult = retryResult;
-      run.retrySuppressedReason = retrySuppressedReason;
-      design.runs.emit(run, 'run_retry_finished', {
-        ...retryAnalyticsBase(eventDecision, failure, errorCode),
-        retry_result: retryResult,
-        ...(retrySuppressedReason
-          ? { retry_suppressed_reason: retrySuppressedReason }
-          : {}),
-      });
-    };
-    let pendingRpcCloseReason = null;
-    const markRpcCloseReason = (reason) => {
-      pendingRpcCloseReason = reason;
-    };
-    const deriveRpcCloseReason = (status, code, signal) => {
-      if (pendingRpcCloseReason) return pendingRpcCloseReason;
-      if (run.cancelRequested || status === 'canceled') return 'cancel_requested';
-      if (signal) return 'signal';
-      if (typeof code === 'number') return code === 0 ? 'exit_0' : 'exit_nonzero';
-      return 'unknown';
-    };
-    const finishWithRetryDecision = (status, code = null, signal = null) => {
-      lifecycle.mark('finalize_start');
-      const result = runResultFromStatus(status);
-      const errorCode = deriveRunErrorCode({
-        status,
-        error: run.error,
-        errorCode: run.errorCode,
-        exitCode: code,
-        signal,
-      });
-      const failure = classifyRunFailure({
-        result,
-        status: {
-          status,
-          error: run.error,
-          errorCode: run.errorCode,
-          exitCode: code,
-          signal,
-        },
-        ...(errorCode ? { errorCode } : {}),
-        agentId: run.agentId,
-        events: run.events,
-      });
-      const sideEffects = {
-        ...scanRunEventsForRetrySideEffects(run.events),
-        cancelRequested: !!run.cancelRequested,
-      };
-      const decision = decideSafeRunRetry({
-        result,
-        failure,
-        attemptCount: run.retryAttemptCount ?? 0,
-        sideEffects,
-      });
-      if (decision.shouldRetry && !design.runs.isTerminal(run.status)) {
-        run.retryAttemptCount = decision.retryAttemptIndex;
-        run.retryFinalResult = undefined;
-        run.retrySuppressedReason = undefined;
-        design.runs.emit(run, 'run_retry_attempted', {
-          ...retryAnalyticsBase(decision, failure, errorCode),
-          retry_reason: decision.retryReason,
-          retry_delay_ms: decision.retryDelayMs,
-        });
-        scheduleRetryRestart(decision.retryDelayMs);
-        return true;
-      }
-      // Resume-on-failure: a terminal *resumable* failure (transient mid-stream
-      // drop / inactivity) on a session-resuming runtime is not a dead end.
-      // Persist the live CLI session so the next turn in this conversation
-      // continues it (`--resume <id>`) instead of opening a fresh session, and
-      // flag the run so the chat can surface a Continue affordance. The session
-      // id is the one we actually drove this attempt with: the resumed id when
-      // continuing, otherwise the freshly minted id we passed via --session-id.
-      //
-      // Gate on a real *committed* boundary this attempt, not merely on bytes
-      // having reached the UI. A completed tool_use / artifact / live-artifact
-      // corresponds to a block the agent has committed to its session (Claude
-      // commits a tool_use block before running the tool), so `--resume` has
-      // something concrete to pick up. We deliberately EXCLUDE
-      // `userVisibleOutputSeen`: it flips true on the first streamed text
-      // delta, but a single-turn drop can stream a few tokens with
-      // `output_tokens == 0` and never commit a text block — resuming that
-      // continues from the prior user turn (nothing to pick up), which is
-      // exactly the "resume something with nothing to continue" case this
-      // feature is meant to avoid. A text-only turn that is cut therefore stays
-      // a from-scratch restart (auto-retry above or a manual Retry).
-      // NOTE: `userVisibleOutputSeen` cannot by itself distinguish "half a text
-      // block, zero commit" from "a committed text block then more streaming";
-      // until the stream exposes a committed-text signal, tool/artifact blocks
-      // are the only reliable resume boundary.
-      const committedWorkSeen = !!(
-        sideEffects.toolCallSeen ||
-        sideEffects.artifactWriteSeen ||
-        sideEffects.liveArtifactSeen
-      );
-      const liveSessionId = agentResumeCtx.isResuming
-        ? agentResumeCtx.resumeSessionId
-        : agentCapturesSessionId
-          ? capturedSessionId
-          : agentResumeCtx.newSessionId;
-      const resumableFailure =
-        result === 'failed' &&
-        def.resumesSessionViaCli === true &&
-        !!run.conversationId &&
-        !!liveSessionId &&
-        committedWorkSeen &&
-        isResumableFailure(failure);
-      run.resumable = resumableFailure;
-      if (resumableFailure) {
-        upsertAgentSession(db, {
-          conversationId: run.conversationId,
-          agentId: def.id,
-          sessionId: liveSessionId,
-          stablePromptHash: currentStableHash,
-          model: safeModel ?? null,
-          cwd: effectiveCwd,
-          lastMessageId: run.assistantMessageId ?? null,
-        });
-      }
-      finalizeRetryTelemetry(status, decision, failure, errorCode);
-      const rpcCloseReason = deriveRpcCloseReason(status, code, signal);
-      design.runs.emit(run, 'diagnostic', {
-        type: 'runtime_close',
-        rpc_close_reason: rpcCloseReason,
-        status,
-        ...(typeof code === 'number' ? { exit_code: code } : {}),
-        ...(signal ? { signal } : {}),
-      });
-      if (executionProfile === 'filesystem' && result === 'success' && visibleAssistantText.trim().length === 0) {
-        const fileNames = filesystemWriteFileNamesFromRunEvents(run.events);
-        if (fileNames.length > 0) {
-          send('agent', {
-            type: 'diagnostic',
-            name: 'filesystem_empty_answer_autofilled',
-            source: 'daemon-run-finalize',
-            fileCount: fileNames.length,
-            files: fileNames.slice(0, 8),
-          });
-          send('agent', {
-            type: 'text_delta',
-            delta: filesystemEmptyAnswerFallbackText(fileNames),
-          });
-        }
-      }
-      pendingRpcCloseReason = null;
-      design.runs.finish(run, status, code, signal);
-      return false;
-    };
-    const mcpServers = buildLiveArtifactsMcpServersForAgent(def, {
-      enabled: Boolean(toolTokenGrant?.token),
-      command: process.execPath,
-      argsPrefix: [OD_BIN],
-    });
-
-    // External MCP servers configured by the user in Settings → External MCP.
-    // Open Design relays them to the agent so the model can call those tools.
-    // Two delivery shapes today:
-    //   - Claude Code: write a `.mcp.json` into the project cwd. Claude Code
-    //     auto-loads that file at spawn (same format the CLI accepts via
-    //     `claude mcp add` + Claude Desktop's config). Fire-and-forget; we
-    //     deliberately do NOT block spawn on a write failure since the agent
-    //     can still run without external tools — log a warning and continue.
-    //   - ACP agents (Hermes/Kimi): merge stdio entries into the existing
-    //     `mcpServers` array; SSE/HTTP entries are skipped because ACP's
-    //     stdio-only descriptor can't represent them yet.
-    // Other agents (Codex, Gemini, OpenCode, Cursor, Qwen, Qoder, Copilot,
-    // Pi, DeepSeek) inherit the user's per-CLI MCP config from their own
-    // home dir for now — a future change can grow this list.
-    //
-    // The MCP config + OAuth tokens were resolved earlier (above
-    // composeDaemonSystemPrompt) so the system prompt could mention any
-    // already-authenticated servers; we reuse `enabledExternalMcp` and
-    // `oauthTokensForSpawn` here for the Claude `.mcp.json` write +
-    // ACP merge so we don't pay for a second filesystem read.
-    //
-    // Claude Code: write `.mcp.json` to the daemon-managed project cwd before
-    // spawn so Claude Code auto-loads the user's external MCP servers. Strict
-    // gating is essential here:
-    //   - cwd must be set (no project → no `.mcp.json` write).
-    //   - cwd must live UNDER PROJECTS_DIR. We never write to a git-linked
-    //     baseDir (= the user's own repo), since that would silently overwrite
-    //     a hand-crafted .mcp.json the user already keeps in their source tree.
-    // We also unlink a stale `.mcp.json` we previously wrote when the user has
-    // since disabled all servers, so removing a server actually takes effect
-    // on the next run.
-    // Dispatch on `def.externalMcpInjection` rather than hard-coding agent
-    // id / stream-format checks. The three branches are functionally
-    // equivalent to the previous shape (claude/acp), with the OpenCode
-    // env-content branch added to fix #2142. Runtimes that leave the field
-    // undefined fall through unchanged — the settings UI surfaces an
-    // explicit "external MCP is not forwarded to <agent>" banner for them
-    // so the previous silent-failure UX is gone.
-    if (
-      def.externalMcpInjection === 'claude-mcp-json' &&
-      isManagedProjectCwd(cwd, PROJECTS_DIR)
-    ) {
-      {
-        const target = path.join(cwd, '.mcp.json');
-        if (enabledExternalMcp.length > 0) {
-          try {
-            const claudeMcp = buildClaudeMcpJson(
-              enabledExternalMcp,
-              oauthTokensForSpawn,
-            );
-            if (claudeMcp) {
-              await fs.promises.mkdir(path.dirname(target), { recursive: true });
-              await fs.promises.writeFile(
-                target,
-                JSON.stringify(claudeMcp, null, 2),
-                'utf8',
-              );
-            }
-          } catch (err) {
-            console.warn(
-              '[mcp-config] failed to write project .mcp.json:',
-              err && err.message ? err.message : err,
-            );
-          }
-        } else {
-          try {
-            await fs.promises.unlink(target);
-          } catch (err) {
-            if ((err && err.code) !== 'ENOENT') {
-              console.warn(
-                '[mcp-config] failed to remove stale .mcp.json:',
-                err && err.message ? err.message : err,
-              );
-            }
-          }
-        }
-      }
-    }
-    if (
-      enabledExternalMcp.length > 0 &&
-      def.externalMcpInjection === 'acp-merge'
-    ) {
-      const acpExternal = buildAcpMcpServers(enabledExternalMcp);
-      mcpServers.push(...acpExternal);
-    }
-    // OpenCode: serialise enabled MCP servers into its `mcp` config schema
-    // and hand the JSON to the child via `OPENCODE_CONFIG_CONTENT`. The env
-    // var is *merged* with the user's saved `~/.config/opencode/opencode
-    // .json` (per OpenCode's documented config layering), so adding a
-    // server here does not erase whatever the user already has in their
-    // global config. We deliberately leave the env unset when no servers
-    // are enabled — overwriting with `{}` would wipe the user's saved
-    // mcp section for this single invocation, which is exactly the kind
-    // of surprise the previous silent-failure UX taught us to avoid.
-    let opencodeConfigContent: string | null = null;
-    const isOpenCodeContent = def.externalMcpInjection === 'opencode-env-content';
-    const isMiMoContent = def.externalMcpInjection === 'mimo-env-content';
-    if (isOpenCodeContent || isMiMoContent) {
-      try {
-        opencodeConfigContent = buildOpenCodeMcpConfigContent(
-          enabledExternalMcp,
-          oauthTokensForSpawn,
-          {
-            allowedDirectories: [effectiveCwd, ...extraAllowedDirs],
-            ...(byokOpenCodeProvider
-              ? { extraConfig: byokOpenCodeProvider.config }
-              : {}),
-          },
-        );
-      } catch (err) {
-        console.warn(
-          '[mcp-config] failed to build OPENCODE_CONFIG_CONTENT:',
-          err && err.message ? err.message : err,
-        );
-      }
-    }
-
-    // Pre-flight the composed prompt against any argv-byte budget the
-    // adapter declared (only DeepSeek TUI today — its CLI doesn't accept
-    // a `-` stdin sentinel, so the prompt has to ride argv). Doing this
-    // before bin resolution means the test harness pins the guard
-    // independently of whether the adapter binary happens to be on PATH
-    // in the CI environment, and the user gets the actionable
-    // adapter-named error even if /api/agents hadn't refreshed yet.
-    const promptBudgetError = checkPromptArgvBudget(def, composed);
-    if (promptBudgetError) {
-      design.runs.emit(
-        run,
-        'error',
-        createSseErrorPayload(
-          promptBudgetError.code,
-          promptBudgetError.message,
-          { retryable: false },
-        ),
-      );
-      return design.runs.finish(run, 'failed', 1, null);
-    }
-
-    let mmdRouteLaunchEnv = null;
-    if (def.id === 'claude' && safeModel) {
-      mmdRouteLaunchEnv = await loadMmdRouteLaunchEnv(
-        {
-          ...process.env,
-          ...(def.env || {}),
-          ...configuredAgentEnv,
-        },
-        safeModel,
-      ).catch(() => null);
-    }
-
-    // agentLaunch / resolvedBin are resolved above the resume guard (hoisted).
-    // Hoisted above the AMR catalog preflight: the empty-catalog branch
-    // below calls `sendAmrAccountFailure(...)` to surface AMR_AUTH_REQUIRED
-    // for signed-out users, and a `const` declared later in the same outer
-    // function scope would hit a TDZ ReferenceError before initialization.
-    const sendAmrAccountFailure = (failure) => {
-      send('error', createSseErrorPayload(
-        failure.code,
-        failure.message,
-        {
-          retryable: false,
-          details: amrAccountFailureDetails(failure),
-        },
-      ));
-    };
-
-    if (def.id === 'amr' && resolvedBin && agentLaunch.launchPath) {
-      const launchPath = agentLaunch.launchPath ?? resolvedBin;
-      const modelProbeEnv = launchPath
-        ? applyAgentLaunchEnv(
-            spawnEnvForAgent(
-              def.id,
-              {
-                ...createAgentRuntimeEnv(process.env, daemonUrl, toolTokenGrant),
-                ...(def.env || {}),
-              },
-              configuredAgentEnv,
-              undefined,
-              { resolvedBin: agentLaunch.selectedPath },
-            ),
-            agentLaunch,
-          )
-        : null;
-      const amrModelScope = resolveAmrProfile(modelProbeEnv ?? process.env);
-      // Resolve the AMR model catalog through the SAME shared cache the UI's
-      // `/api/amr/models` endpoint serves (AmrModelLoadingCache): a cached
-      // authoritative `vela model list` when it is hot, otherwise the offline
-      // `vela model preset` seed while a remote refresh runs in the background.
-      //
-      // Why not a fresh `vela model list` per run: that authoritative call
-      // needs network reachability to the AMR gateway AND `$HOME` (the offline
-      // `preset`/`--version` calls need neither), takes up to ~10s, and only
-      // retries a narrow set of network errors. Running it blocking on every
-      // turn turned any transient gateway/timeout/HOME hiccup into a hard
-      // "AMR model … is not available from Vela" — even for a logged-in user
-      // who already picked a real model the picker surfaced from the preset
-      // seed. Under CorpLink/飞连 the call routinely exceeded the timeout, so
-      // AMR became unusable in packaged nightlies. Reusing the cache keeps that
-      // blocking probe off the per-run hot path and degrades to preset instead
-      // of fail-closing; vela's own `session/set_model` remains the final gate.
-      let liveModels = [];
-      try {
-        const probe = await resolveAmrModelProbe({ dataDir: RUNTIME_DATA_DIR, env: process.env, readAppConfig });
-        const catalog = await amrModelLoadingCache.get(probe.cacheKey, {
-          fetchPreset: () => fetchVelaPresetModels(probe.launchPath, probe.env),
-          fetchRemote: () => fetchVelaRemoteModelsWithRetry(probe.launchPath, probe.env),
-        });
-        liveModels = catalog.models ?? [];
-      } catch (error) {
-        // Do not swallow silently: a probe failure here is exactly what made
-        // the packaged AMR breakage undiagnosable (the old `catch {}` left no
-        // trace in any log or diagnostics bundle). Record it and degrade to the
-        // remembered catalog below.
-        console.warn('[amr] model catalog preflight probe failed', error);
-        liveModels = [];
-      }
-      const rememberedLiveModels = getRememberedLiveModels(def.id, amrModelScope);
-      if (liveModels.length > 0) {
-        rememberLiveModels(def.id, liveModels, amrModelScope);
-      }
-      liveModels = preferFreshLiveModels(liveModels, rememberedLiveModels);
-      const liveModelIds = new Set(
-        liveModels.map((candidate) => candidate?.id).filter(Boolean),
-      );
-      // A request that came in as 'default'/empty is normally pre-resolved to a
-      // concrete id via the agent-wide cached model order; if it still is not,
-      // adopt the catalog's enabled default so the spawn layer always has a
-      // usable real id.
-      const userAskedForDefault =
-        typeof model !== 'string' ||
-        !model.trim() ||
-        model.trim().toLowerCase() === 'default';
-      const defaultRunModel = resolveDefaultModelFromOptions(liveModels);
-      if (
-        !safeModel ||
-        safeModel === 'default' ||
-        (
-          userAskedForDefault &&
-          !hasDefaultModelEnvOverride &&
-          defaultRunModel &&
-          (!liveModelIds.has(safeModel) || safeModel !== defaultRunModel)
-        )
-      ) {
-        safeModel = defaultRunModel ?? (safeModel === 'default' ? null : safeModel ?? null);
-        agentOptions.model = safeModel;
-      }
-      if (liveModelIds.size === 0) {
-        // The catalog is genuinely empty: even the offline preset seed could
-        // not be read, which almost always means the user is signed out (`vela`
-        // catalog calls 401) or the CLI is unrunnable. Prefer the relogin
-        // affordance over a misleading "choose a model".
-        if (def.id === 'amr') {
-          const loginStatus = readVelaLoginStatus(
-            modelProbeEnv ?? process.env,
-            configuredAgentEnv,
-          );
-          if (!loginStatus.loggedIn) {
-            sendAmrAccountFailure({
-              code: 'AMR_AUTH_REQUIRED',
-              message:
-                'AMR sign-in is required. Sign in to AMR Cloud again, then retry this run.',
-              action: 'relogin',
-            });
-            return design.runs.finish(run, 'failed', 1, null);
-          }
-        }
-        // Logged in but no catalog at all AND no resolvable model: only now is
-        // there nothing safe to forward, so surface the model error.
-        if (!safeModel) {
-          send('error', createAmrModelUnavailablePayload(safeModel, {
-            reason: 'model_catalog_unavailable',
-          }));
-          return design.runs.finish(run, 'failed', 1, null);
-        }
-        // Otherwise fall through with the user's selected model and let vela's
-        // `session/set_model` be the authoritative gate.
-      } else if (!safeModel) {
-        // Catalog known but we could not resolve any model id to forward.
-        send('error', createAmrModelUnavailablePayload(
-          typeof model === 'string' && model.trim() ? model : safeModel,
-          { availableModels: [...liveModelIds] },
-        ));
-        return design.runs.finish(run, 'failed', 1, null);
-      }
-      // NOTE: when the selected model is absent from the (possibly preset-only
-      // or stale) catalog we intentionally do NOT fail-close. The cached/preset
-      // catalog can lag the live one, and a logged-in user picked a concrete
-      // id; vela rejects a truly unsupported model at `session/set_model` with
-      // a precise error, which beats a pre-emptive block on a flaky metadata read.
-    }
-
-    // Plain-streaming adapters that own a "continue most recent
-    // conversation" CLI flag (today: only `agy -c`) read this signal
-    // to resume upstream session state on follow-up turns. The query
-    // matches any persisted assistant message in the same conversation
-    // EXCEPT the placeholder row this run just inserted (it's still
-    // `pending` and has no body — counting it as prior would always
-    // force `-c` on the very first turn). Adapters that don't consume
-    // this field ignore it.
-    const hasPriorAssistantTurn = run.conversationId
-      ? Boolean(
-          db
-            .prepare(
-              `SELECT 1 FROM messages
-               WHERE conversation_id = ?
-                 AND role = 'assistant'
-                 AND COALESCE(content, '') <> ''
-                 AND id <> COALESCE(?, '')
-               LIMIT 1`,
-            )
-            .get(run.conversationId, run.assistantMessageId ?? ''),
-        )
-      : false;
-
-    // Antigravity's `agy` is silent on stdout/stderr in print mode for
-    // both auth-missing and quota-exhausted failures — the actual
-    // RESOURCE_EXHAUSTED / "not logged in" payload only surfaces in
-    // its `--log-file`. We allocate a per-run temp path, pipe agy's
-    // log to it via buildArgs, then read it in the empty-output guard
-    // to disambiguate the silent-failure cause. Other adapters ignore
-    // this field.
-    const agentLogFilePath =
-      def.id === 'antigravity'
-        ? path.join(os.tmpdir(), `od-agy-${run.id}.log`)
-        : undefined;
-    const promptFile = await preparePromptFileForAgent(def, composed, run.id);
-    const cleanupPromptFile = () => {
-      if (promptFile) promptFile.cleanup().catch(() => {});
-    };
-
-    // Codex CLI parses config.toml before processing any -c overrides. An
-    // invalid `service_tier` value (the Codex app has written "priority",
-    // "default", and other values the CLI rejects) causes an immediate parse
-    // error and exit-1 before any work starts. Normalize it in-place — any
-    // value outside {fast,flex} has its line removed so the CLI uses its
-    // built-in default — so the launch succeeds. Errors are silently swallowed
-    // — a missing or read-only config.toml is fine, and the Codex CLI still
-    // surfaces the original error if the write fails. See issue #4276 / #3408.
-    if (def.id === 'codex') {
-      const { normalizeCodexConfigFile } = await import('./codex-config-normalize.js');
-      // Route through spawnEnvForAgent so resolveCodexConfigPath sees the same
-      // fully-expanded CODEX_HOME the Codex child process will see. In
-      // particular, spawnEnvForAgent calls expandConfiguredEnv which expands
-      // `~/` / `~\` prefixes — a user-configured CODEX_HOME="~/.codex-alt"
-      // would otherwise resolve to the literal path "~/.codex-alt/config.toml"
-      // in the normalizer while the child resolves it to the absolute path,
-      // leaving the real config untouched. Mirrors the diagnostics-export.ts
-      // `envFor('codex')` pattern. See issue #4276.
-      await normalizeCodexConfigFile(
-        spawnEnvForAgent('codex', process.env, configuredAgentEnv),
-      );
-    }
-
-    // Serialize antigravity spawns whose buildArgs writes a concrete
-    // model into settings.json. Two concurrent runs with different
-    // models would otherwise race the file: A writes model A, B writes
-    // model B, then A's agy reads model B. The lock is acquired BEFORE
-    // buildArgs (which performs the write) and released asynchronously
-    // AFTER agy's --log-file confirms the model was propagated. See
-    // `antigravity.ts` for the chain implementation.
-    let antigravityModelLockRelease: (() => void) | null = null;
-    const antigravityConcreteModel =
-      def.id === 'antigravity'
-      && typeof agentOptions.model === 'string'
-      && agentOptions.model.length > 0
-      && agentOptions.model !== 'default'
-        ? agentOptions.model
-        : null;
-    if (antigravityConcreteModel) {
-      const { acquireAntigravityModelLock } = await import(
-        './runtimes/defs/antigravity.js'
-      );
-      antigravityModelLockRelease = await acquireAntigravityModelLock();
-    }
-
-    let args;
-    try {
-      args = def.buildArgs(
-        composed,
-        safeImages,
-        extraAllowedDirs,
-        agentOptions,
-        {
-          cwd: effectiveCwd,
-          hasPriorAssistantTurn,
-          agentLogFilePath,
-          promptFilePath: promptFile?.path,
-          resumeSessionId: agentResumeCtx.resumeSessionId,
-          newSessionId: agentResumeCtx.newSessionId,
-        },
-      );
-    } catch (err) {
-      cleanupPromptFile();
-      throw err;
-    }
-    // Second-pass budget check that knows about the Windows `.cmd` shim
-    // wrap. The pre-buildArgs `checkPromptArgvBudget` only looks at the
-    // raw composed prompt; on Windows an npm-installed adapter resolves
-    // to e.g. `deepseek.cmd`, the spawn path goes through `cmd.exe /d /s
-    // /c "<inner>"`, and `quoteForWindowsCmdShim` doubles every embedded
-    // `"` plus wraps any whitespace/special-char arg in outer quotes —
-    // so a quote-heavy prompt that fit under `maxPromptArgBytes` can
-    // still expand past CreateProcess's 32_767-char cap. Fail fast with
-    // the same `AGENT_PROMPT_TOO_LARGE` shape so the SSE error path
-    // doesn't have to special-case it.
-    const cmdShimBudgetError = checkWindowsCmdShimCommandLineBudget(
-      def,
-      agentLaunch.launchPath ?? resolvedBin,
-      args,
-    );
-    if (cmdShimBudgetError) {
-      cleanupPromptFile();
-      design.runs.emit(
-        run,
-        'error',
-        createSseErrorPayload(
-          cmdShimBudgetError.code,
-          cmdShimBudgetError.message,
-          { retryable: false },
-        ),
-      );
-      return design.runs.finish(run, 'failed', 1, null);
-    }
-
-    // Companion guard for non-shim Windows installs (e.g. a cargo-built
-    // `deepseek.exe` rather than the npm `.cmd` shim). Direct `.exe`
-    // spawns skip the cmd.exe wrap above, but Node/libuv still composes
-    // a CreateProcess `lpCommandLine` by walking each argv element
-    // through `quote_cmd_arg`, which escapes every embedded `"` as `\"`
-    // and doubles backslashes adjacent to quotes. A quote-heavy prompt
-    // under `maxPromptArgBytes` can expand past the 32_767-char kernel
-    // cap there too, so the cmd-shim early-return alone would let those
-    // users hit a generic `spawn ENAMETOOLONG`.
-    const directExeBudgetError = checkWindowsDirectExeCommandLineBudget(
-      def,
-      agentLaunch.launchPath ?? resolvedBin,
-      args,
-    );
-    if (directExeBudgetError) {
-      cleanupPromptFile();
-      design.runs.emit(
-        run,
-        'error',
-        createSseErrorPayload(
-          directExeBudgetError.code,
-          directExeBudgetError.message,
-          { retryable: false },
-        ),
-      );
-      return design.runs.finish(run, 'failed', 1, null);
-    }
-
-    let persistDeliveredAgentSessionState = () => {};
-    if (def.resumesSessionViaCli === true && run.conversationId) {
-      let persisted = false;
-      persistDeliveredAgentSessionState = () => {
-        if (persisted) return;
-        persisted = true;
-        // The id to persist for a create turn: capture-style adapters store the
-        // session id the CLI minted and reported on the stream; specify-style
-        // adapters store the daemon-minted id they passed to the CLI. A
-        // capture-style run that never reported an id (CLI died before
-        // `thread.started`) leaves nothing to resume — correct, the next turn
-        // starts fresh and re-seeds the transcript.
-        const createTurnSessionId = agentCapturesSessionId
-          ? capturedSessionId
-          : agentResumeCtx.newSessionId;
-        if (!agentResumeCtx.isResuming && createTurnSessionId) {
-          upsertAgentSession(db, {
-            conversationId: run.conversationId,
-            agentId: def.id,
-            sessionId: createTurnSessionId,
-            stablePromptHash: currentStableHash,
-            model: safeModel ?? null,
-            cwd: effectiveCwd,
-            lastMessageId: run.assistantMessageId ?? null,
-          });
-          return;
-        }
-        if (agentResumeCtx.isResuming && agentResumeCtx.resumeSessionId) {
-          // Advance the resume identity guard after a successful resume turn:
-          // the conversation grew by this turn, so the cursor must move to the
-          // new max position (otherwise the next turn sees `cursor + 4` and
-          // falsely reseeds). model/cwd are unchanged (they matched on resume);
-          // refresh the stable hash to what the session now holds.
-          upsertAgentSession(db, {
-            conversationId: run.conversationId,
-            agentId: def.id,
-            sessionId: agentResumeCtx.resumeSessionId,
-            stablePromptHash: currentStableHash,
-            model: safeModel ?? null,
-            cwd: effectiveCwd,
-            lastMessageId: run.assistantMessageId ?? null,
-          });
-        }
-      };
-    }
-
-    // `runStartTimeMs` is consumed by the run-end artifact-manifest
-    // reconciler (#2893 / #3110) to skip artifacts whose mtime predates
-    // this run. The original main-side hunk also re-declared `const send`
-    // here; on this branch `send` was hoisted into the AMR preflight
-    // earlier, so we keep only the new `runStartTimeMs` declaration.
-    const runStartTimeMs = Date.now();
-    const inactivityTimeoutMs = resolveChatRunInactivityTimeoutMs(def.inactivityTimeoutMs);
-    const artifactQuietPeriodMs = resolveChatRunArtifactQuietPeriodMs();
-    // Grace before the inactivity watchdog escalates a stalled child from
-    // SIGTERM to SIGKILL. Env-tunable like its OD_CHAT_RUN_* cancel-grace
-    // siblings so the escalation path can be exercised deterministically.
-    const inactivityKillGraceMs = (() => {
-      const raw = Number(process.env.OD_CHAT_RUN_INACTIVITY_KILL_GRACE_MS);
-      return Number.isFinite(raw) && raw > 0 ? raw : 3_000;
-    })();
-    let inactivityTimer = null;
-    let childStdoutSeen = false;
-    let lastAgentEventPhase = 'spawn pending';
-    let lastToolResultChars = 0;
-    // Becomes true once any live-artifact create has been registered for
-    // this run. Subsequent watchdog scheduling uses the shorter quiet
-    // period, and a watchdog trip after this point is treated as
-    // "agent finished the deliverable and went idle" rather than
-    // "agent stalled with nothing to show" (issue #1451).
-    let artifactRegistered = false;
-    // Only daemon-initiated quiet-period termination should be treated
-    // as `succeeded` in the close handler. A later unrelated SIGTERM /
-    // SIGKILL (external `kill`, OOM, container shutdown) must keep its
-    // existing `failed` classification even when `artifactRegistered`
-    // is true — those signals don't mean the agent finished cleanly,
-    // they just terminated the process. Set strictly inside
-    // `failForInactivity`'s quiet-period branch.
-    let artifactQuietShutdownRequested = false;
-    // Set when the no-output inactivity watchdog routed this attempt through
-    // the same-run retry finalizer AND that finalizer restarted the run on a
-    // fresh child. The stalled child is then SIGTERM'd, so its later `close`
-    // must NOT finalize the run a second time or unregister the new attempt's
-    // event sink / run handle (both keyed by the shared runId). The close
-    // handler bails early when this is true, revoking only this attempt's own
-    // tool token.
-    let watchdogRetryRestarted = false;
-    const summarizeAgentEventForInactivity = (payload) => {
-      const type = payload?.type ? String(payload.type) : 'unknown';
-      if (type === 'tool_result') {
-        const content = typeof payload.content === 'string' ? payload.content : '';
-        lastToolResultChars = Math.max(lastToolResultChars, content.length);
-        return `tool_result:${content.length} chars`;
-      }
-      if (type === 'tool_use') {
-        const name = payload?.name ? String(payload.name) : 'unknown';
-        return `tool_use:${name}`;
-      }
-      if (type === 'text_delta' || type === 'thinking_delta') {
-        const text = typeof payload.delta === 'string'
-          ? payload.delta
-          : typeof payload.text === 'string'
-            ? payload.text
-            : '';
-        return `${type}:${text.length} chars`;
-      }
-      if (type === 'status') {
-        const label = payload?.label ? String(payload.label) : 'unknown';
-        return `status:${label}`;
-      }
-      return type;
-    };
-    const clearInactivityWatchdog = () => {
-      if (inactivityTimer) {
-        clearTimeout(inactivityTimer);
-        inactivityTimer = null;
-      }
-    };
-    let forcedChildShutdownTimers = [];
-    const clearForcedChildShutdown = () => {
-      for (const timer of forcedChildShutdownTimers) clearTimeout(timer);
-      forcedChildShutdownTimers = [];
-    };
-    const scheduleForcedChildShutdown = () => {
-      if (!child) return;
-      clearForcedChildShutdown();
-      // Capture THIS attempt's child and its process group. A same-run retry
-      // can swap `run.child` to a fresh child within the grace window; these
-      // timers must escalate the stalled child they were scheduled for, never
-      // whatever now occupies `run.child` — otherwise the healthy retry gets
-      // killed and this stalled child is left unreaped. See runs.ts
-      // `signalChildProcess`.
-      const targetChild = child;
-      const targetProcessGroupId = run.processGroupId;
-      forcedChildShutdownTimers = [
-        setTimeout(() => {
-          design.runs.signalChildProcess(targetChild, targetProcessGroupId, 'SIGTERM');
-        }, inactivityKillGraceMs),
-        setTimeout(() => {
-          design.runs.signalChildProcess(targetChild, targetProcessGroupId, 'SIGKILL');
-        }, inactivityKillGraceMs * 2),
-      ];
-    };
-    const failForInactivity = () => {
-      if (run.cancelRequested || design.runs.isTerminal(run.status)) return;
-      clearInactivityWatchdog();
-      if (artifactRegistered) {
-        // The deliverable already exists. The agent process is either
-        // genuinely idle (claude-code's stream-json child sitting on an
-        // open stdin) or wedged in post-write reasoning that never
-        // emits stdout. Either way, finishing the run via the normal
-        // child-exit path (status decision in child.on('close') below)
-        // is safer than tearing it down with a failure banner — the
-        // tool token, cancel state, and exit-code classification stay
-        // owned by the existing lifecycle. SIGTERM the child and let
-        // the close handler classify the run as succeeded (via the
-        // artifactQuietShutdown branch). Mark this termination as
-        // daemon-initiated so an unrelated later signal (external
-        // kill, OOM) is NOT silently reclassified to `succeeded` —
-        // only signals from this watchdog branch should be.
-        artifactQuietShutdownRequested = true;
-        if (acpSession?.abort) {
-          acpSession.abort();
-        }
-        if (child && !child.killed) design.runs.signalChild(run, 'SIGTERM');
-        scheduleForcedChildShutdown();
-        return;
-      }
-      // OpenCode retries a 429 usage-limit silently and emits nothing on
-      // stdout/stderr, so the watchdog is the first signal we get. The real
-      // reason is recorded only in OpenCode's own session log — recover it
-      // and surface it HERE, before finish() tears down the live SSE
-      // clients, so a viewer sees "usage limit reached" instead of the
-      // generic stall message. Bound to this run via `since` so a stale or
-      // concurrent session's error can't be misattributed. See issue #982.
-      let stallPayload = null;
-      if (agentId === 'opencode') {
-        const logFailure = readOpenCodeServiceFailure(spawnedAgentEnv, {
-          since: run.createdAt,
-        });
-        if (logFailure) {
-          stallPayload = createSseErrorPayload(
-            logFailure.code,
-            logFailure.message,
-            { retryable: logFailure.retryable },
-          );
-        }
-      }
-      if (!stallPayload) {
-        const message =
-          `Agent stalled without emitting any new output for ${Math.round(inactivityTimeoutMs / 1000)}s. ` +
-          'The model or CLI likely hung while generating. ' +
-          `Phase details: spawned agent ${userFacingAgentLabel(agentId, resolvedBin)}; stdout arrived: ${childStdoutSeen ? 'yes' : 'no'}; ` +
-          `last agent event: ${lastAgentEventPhase}; largest tool result observed: ${lastToolResultChars} chars. ` +
-          'Retry the turn, pick a different model, or start a new conversation if the prior context is very large.';
-        stallPayload = createSseErrorPayload('AGENT_EXECUTION_FAILED', message, { retryable: true });
-      }
-      send('error', stallPayload);
-      // A silent first-token hang is one of the safe transient failure shapes
-      // this run is allowed to recover: classifyRunFailure maps the stall text
-      // to a retryable `timeout` at `first_token_wait`, and decideSafeRunRetry
-      // permits the same-run retry when no output/tools/artifacts were seen.
-      // Route through the shared finalizer (after surfacing stallPayload) so
-      // the watchdog path gets the same run_retry_attempted/run_retry_finished
-      // telemetry as child close/error — not a bare terminal failure.
-      const retried = finishWithRetryDecision('failed', 1, null);
-      if (retried) {
-        watchdogRetryRestarted = true;
-      }
-      if (acpSession?.abort) {
-        acpSession.abort();
-      }
-      if (child && !child.killed) design.runs.signalChild(run, 'SIGTERM');
-      scheduleForcedChildShutdown();
-    };
-    const activeInactivityTimeoutMs = () =>
-      resolveActiveInactivityTimeoutMs({
-        inactivityTimeoutMs,
-        artifactQuietPeriodMs,
-        artifactRegistered,
-      });
-    const noteAgentActivity = () => {
-      const delay = activeInactivityTimeoutMs();
-      if (delay <= 0) return;
-      clearInactivityWatchdog();
-      inactivityTimer = setTimeout(failForInactivity, delay);
-      inactivityTimer.unref?.();
-    };
-    const noteArtifactRegistered = () => {
-      if (artifactRegistered) return;
-      artifactRegistered = true;
-      // Switch the watchdog to the shorter quiet-period window
-      // immediately so we don't have to wait for the next agent event
-      // before the new ceiling takes effect. Call unconditionally:
-      // an earlier `if (inactivityTimer)` gate left the run in limbo
-      // when `OD_CHAT_RUN_INACTIVITY_TIMEOUT_MS=0` but
-      // `OD_CHAT_RUN_ARTIFACT_QUIET_PERIOD_MS>0` — noteAgentActivity()
-      // had returned early at run start (pre-artifact delay = 0,
-      // no timer set), so the guard then skipped the re-arm and the
-      // newly-positive quiet-period delay never armed a timer at all.
-      // `noteAgentActivity` itself is the one that decides whether to
-      // schedule (it bails when the active delay is 0), so leaving the
-      // decision there keeps the behavior coherent across all four
-      // combinations of pre / quiet timeouts.
-      noteAgentActivity();
-    };
-    const unregisterChatAgentEventSink = () => {
-      const sinkRunId = toolTokenGrant?.runId ?? runId;
-      activeChatAgentEventSinks.delete(sinkRunId);
-      activeChatRunHandles.delete(sinkRunId);
-    };
-    if (toolTokenGrant?.runId) {
-      activeChatAgentEventSinks.set(toolTokenGrant.runId, (payload) => {
-        lastAgentEventPhase = summarizeAgentEventForInactivity(payload);
-        noteAgentActivity();
-        send('agent', payload);
-      });
-      activeChatRunHandles.set(toolTokenGrant.runId, { noteArtifactRegistered });
-    }
-    // If detection can't find the binary, surface a friendly SSE error
-    // pointing at /api/agents instead of silently falling back to
-    // spawn(def.bin) — that fallback re-introduces the exact ENOENT symptom
-    // from issue #10.
-    if (!resolvedBin || !agentLaunch.launchPath) {
-      cleanupPromptFile();
-      revokeToolToken('child_exit');
-      unregisterChatAgentEventSink();
-      send('error', createSseErrorPayload(
-        'AGENT_UNAVAILABLE',
-        `Agent "${def.name}" (\`${def.bin}\`) is not installed or not on PATH. ` +
-          'Install it and refresh the agent list (GET /api/agents) before retrying.',
-        { retryable: true },
-      ));
-      return design.runs.finish(run, 'failed', 1, null);
-    }
-    const browserUseRuntimeEnv = run.browserUse
-      ? {
-          OD_BROWSER_USE_REQUESTED: run.browserUse.requested ? '1' : '0',
-          OD_BROWSER_USE_AVAILABLE: run.browserUse.available ? '1' : '0',
-          ...(run.browserUse.reason ? { OD_BROWSER_USE_UNAVAILABLE_REASON: run.browserUse.reason } : {}),
-          OD_BROWSER_USE_REGISTRY_PATH: run.browserUse.diagnostics?.registryPath ?? '',
-        }
-      : {};
-    const agentSpawnEnv = spawnEnvForAgent(
-      def.id,
-      {
-        ...createAgentRuntimeEnv(process.env, daemonUrl, toolTokenGrant),
-        ...(def.env || {}),
-        ...browserUseRuntimeEnv,
-      },
-      configuredAgentEnv,
-      undefined,
-      { resolvedBin: agentLaunch.selectedPath },
-    );
-    if (def.id === 'amr') {
-      const loginStatus = readVelaLoginStatus(agentSpawnEnv, configuredAgentEnv);
-      if (!loginStatus.loggedIn) {
-        cleanupPromptFile();
-        revokeToolToken('child_exit');
-        unregisterChatAgentEventSink();
-        sendAmrAccountFailure({
-          code: 'AMR_AUTH_REQUIRED',
-          message: 'AMR sign-in is required. Sign in to AMR Cloud again, then retry this run.',
-          action: 'relogin',
-        });
-        return design.runs.finish(run, 'failed', 1, null);
-      }
-    }
-    const odMediaEnv = {
-      OD_BIN,
-      OD_NODE_BIN,
-      OD_DAEMON_URL: daemonUrl,
-      ...(typeof projectId === 'string' && projectId && cwd
-        ? {
-            OD_PROJECT_ID: projectId,
-            OD_PROJECT_DIR: cwd,
-          }
-        : {}),
-    };
-    if (run.cancelRequested || design.runs.isTerminal(run.status)) {
-      cleanupPromptFile();
-      revokeToolToken('child_exit');
-      unregisterChatAgentEventSink();
-      return;
-    }
-
-    run.status = 'running';
-    run.updatedAt = Date.now();
-    send('start', {
-      runId,
-      agentId,
-      bin: userFacingAgentLabel(agentId, resolvedBin),
-      streamFormat: def.streamFormat ?? 'plain',
-      projectId: typeof projectId === 'string' ? projectId : null,
-      cwd,
-      model: safeModel,
-      reasoning: safeReasoning,
-      toolTokenExpiresAt: toolTokenGrant?.expiresAt ?? null,
-    });
-    noteAgentActivity();
-
-    let child;
-    let acpSession = null;
-    let writePromptToChildStdin = false;
-    let spawnedAgentEnv = null;
-    let agentStdoutTail = '';
-    let agentStderrTail = '';
-    const agentStderrFilter = createAgentStderrVisibilityFilter(agentId);
-    const emitVisibleAgentStderr = (chunk: unknown) => {
-      const visibleChunk = agentStderrFilter.write(chunk);
-      if (!visibleChunk) return;
-      agentStderrTail = `${agentStderrTail}${visibleChunk}`.slice(-2000);
-      send('stderr', { chunk: visibleChunk });
-    };
-    const flushVisibleAgentStderr = () => {
-      const visibleChunk = agentStderrFilter.flush();
-      if (!visibleChunk) return;
-      agentStderrTail = `${agentStderrTail}${visibleChunk}`.slice(-2000);
-      send('stderr', { chunk: visibleChunk });
-    };
-    try {
-      // Prompt delivery via stdin is now the universal default. This bypasses
-      // both the cmd.exe 8KB limit and the CreateProcess 32KB limit.
-      const stdinMode =
-        def.promptViaStdin || def.streamFormat === 'acp-json-rpc'
-          ? 'pipe'
-          : 'ignore';
-      const env = applyAgentLaunchEnv({
-        ...agentSpawnEnv,
-        ...(mmdRouteLaunchEnv || {}),
-        ...odMediaEnv,
-        ...(byokOpenCodeProvider ? byokOpenCodeProvider.env : {}),
-        ...openDesignAmrTraceEnv({
-          agentId: def.id,
-          runId: run.id,
-          conversationId: run.conversationId,
-          runAttempt: run.retryAttemptCount ?? 0,
-        }),
-        // OpenCode external-MCP injection (issue #2142). Layered AFTER
-        // spawnEnvForAgent / odMediaEnv / configuredAgentEnv so the
-        // daemon-built MCP config wins over a stale value the user
-        // might have exported in their shell — that would let an
-        // outdated content string suppress the user's freshly-saved
-        // MCP servers, which is exactly the bug we are fixing.
-        // `opencodeConfigContent === null` means "no enabled servers";
-        // we deliberately leave the env unset in that case so the
-        // user's saved `~/.config/opencode/opencode.json` continues
-        // to apply as-is.
-        ...(opencodeConfigContent
-          ? { [isMiMoContent ? 'MIMOCODE_CONFIG_CONTENT' : 'OPENCODE_CONFIG_CONTENT']: opencodeConfigContent }
-          : {}),
-      }, agentLaunch);
-      spawnedAgentEnv = env;
-      const invocation = createCommandInvocation({
-        command: agentLaunch.launchPath,
-        args,
-        env,
-      });
-      lifecycle.mark('launch_preflight_end');
-      lifecycle.mark('process_spawn_start');
-      child = spawn(invocation.command, invocation.args, {
-        env,
-        stdio: [stdinMode, 'pipe', 'pipe'],
-        cwd: effectiveCwd,
-        shell: false,
-        detached: process.platform !== 'win32',
-        // Required when invocation wraps a Windows .cmd/.bat shim through
-        // cmd.exe; without this, Node re-escapes the inner command line and
-        // breaks paths containing spaces (issue #315).
-        windowsVerbatimArguments: invocation.windowsVerbatimArguments,
-      });
-      lifecycle.mark('process_spawned');
-      run.child = child;
-      run.childPid = typeof child.pid === 'number' ? child.pid : null;
-      run.processGroupId =
-        process.platform !== 'win32' && typeof child.pid === 'number'
-          ? child.pid
-          : null;
-      // Schedule release of the antigravity model lock once agy's
-      // --log-file confirms the chosen model was propagated to the
-      // backend (the upstream signal that settings.json was read).
-      // The watcher's `false` return (timeout) deliberately does NOT
-      // release — looper review at 263fd2fe7 flagged that releasing
-      // on timeout reopens the slow-cold-start race: a >15s agy
-      // startup that hadn't yet read settings.json would let run B
-      // rewrite the file and run A would then read run B's model.
-      // The exit handler is the canonical fallback that releases the
-      // lock no matter what (crashed agy, fast exit, etc.) so the
-      // queue can never starve permanently.
-      if (
-        antigravityModelLockRelease
-        && antigravityConcreteModel
-        && agentLogFilePath
-      ) {
-        const releaseOnce = (() => {
-          let fired = false;
-          return () => {
-            if (fired) return;
-            fired = true;
-            antigravityModelLockRelease?.();
-          };
-        })();
-        const watcherAbort = new AbortController();
-        const { waitForAgyToReadModel } = await import(
-          './runtimes/defs/antigravity.js'
-        );
-        void waitForAgyToReadModel(
-          agentLogFilePath,
-          antigravityConcreteModel,
-          { abortSignal: watcherAbort.signal },
-        )
-          .then((found) => {
-            // Only release on TRUE confirmation; a `false` return means
-            // the watcher ran out of its polling window without seeing
-            // the propagation line. We hold the lock until child exit
-            // so a slow-cold-start agy can't be pre-empted by a
-            // concurrent settings.json rewrite from run B.
-            if (found) releaseOnce();
-          })
-          .catch(() => undefined);
-        child.once('exit', () => {
-          // Stop the watcher so its pending readFile / setTimeout
-          // chain does not outlive the run and leak into subsequent
-          // antigravity spawns (or test cases).
-          watcherAbort.abort();
-          releaseOnce();
-        });
-      }
-      if (def.promptViaStdin && child.stdin && def.streamFormat !== 'pi-rpc') {
-        // EPIPE from a fast-exiting CLI (bad auth, missing model, exit on
-        // launch) would otherwise surface as an unhandled stream error and
-        // crash the daemon. Swallow it — the regular exit/close handlers
-        // below already route the underlying failure to SSE via stderr.
-        child.stdin.on('error', (err) => {
-          // EPIPE = Unix broken-pipe when child closes its stdin read end
-          // early. 'write EOF' (err.code 'EOF') = Windows equivalent of
-          // the same condition via UV_EOF. Both mean the child exited before
-          // reading stdin — the process exit/close handlers already route
-          // the underlying failure to SSE via stderr, so swallow these here.
-          if (err.code !== 'EPIPE' && err.code !== 'EOF' && err.message !== 'write EOF') {
-            send(
-              'error',
-              createSseErrorPayload(
-                'AGENT_EXECUTION_FAILED',
-                `stdin: ${err.message}`,
-              ),
-            );
-          }
-        });
-        writePromptToChildStdin = true;
-      }
-    } catch (err) {
-      cleanupPromptFile();
-      revokeToolToken('child_exit');
-      unregisterChatAgentEventSink();
-      send('error', createSseErrorPayload('AGENT_EXECUTION_FAILED', `spawn failed: ${err.message}`));
-      design.runs.finish(run, 'failed', 1, null);
-      return;
-    }
-
-    child.stdout.setEncoding('utf8');
-    child.stderr.setEncoding('utf8');
-
-    // Reset the inactivity watchdog on every raw stdout byte so that
-    // structured adapters that buffer partial lines (Codex item.completed,
-    // pi-rpc session/prompt, ACP agent messages) and models that spend a
-    // long time in non-streamed reasoning still keep the run alive.
-    child.stdout.on('data', (chunk) => {
-      childStdoutSeen = true;
-      noteAgentActivity();
-      agentStdoutTail = `${agentStdoutTail}${chunk}`.slice(-2000);
-    });
-
-    // ---- Memory: assistant-reply capture for LLM extraction --------------
-    // Hand the extractor the guarded, rendered reply (`memoryReplyText`, fed
-    // through `send()` from either the `agent` text_delta or the `stdout`
-    // channel), NOT the child's raw stdout. For stream-json agents (Claude Code)
-    // raw stdout is JSONL transport — system:init, stream_event thinking deltas,
-    // hook_started/hook_response frames — none of which is the reply; mining it
-    // produced empty extractions that, near-identical across a build's re-fires,
-    // caused the same turn to be re-analyzed dozens of times.
-    child.on('close', () => {
-      const userMsg = typeof message === 'string' ? message : '';
-      // Forward the chat agent id so memory-llm.pickProvider can
-      // constrain its auto-pick to the chat protocol's family — keeps
-      // a Claude Code (anthropic) chat from triggering OpenAI/gpt-4o-
-      // mini extraction in the background just because the user has
-      // an OpenAI key parked in media-config.
-      const memoryOptions = {
-        projectRoot: PROJECT_ROOT,
-        chatAgentId: typeof agentId === 'string' ? agentId : null,
-        chatModel: typeof safeModel === 'string' ? safeModel : null,
-        // Scope the extractor's duplicate-turn de-dup to this conversation, so a
-        // re-fired turn collapses but an identical (message, reply) in another
-        // conversation is still examined.
-        conversationId: run.conversationId ?? null,
-      };
-      void import('./memory-llm.js')
-        .then(({ extractWithLLM, distillAnnotationsToMemory }) => {
-          // Read the reply HERE, in the post-import microtask, not in the
-          // synchronous close handler: the Claude stream flush is a later
-          // 'close' listener, so deferring the read lets flush() emit the reply's
-          // final buffered frame first and a reply that ends without a trailing
-          // newline isn't truncated.
-          const captured = memoryReplyText;
-          const generalPass = extractWithLLM(
-            RUNTIME_DATA_DIR,
-            {
-              userMessage: userMsg,
-              assistantMessage: captured,
-            },
-            memoryOptions,
-          );
-          // Auto-distill any inline preview feedback (comments / highlights /
-          // drawn marks) from this turn into durable feedback + rule memory.
-          // This closes the "interaction → memory" loop automatically: the
-          // agent no longer has to propose a rule and the user no longer has
-          // to click Keep — a review turn that carried annotations mines
-          // itself in the background and writes straight to the store.
-          const annotationPass =
-            safeCommentAttachments.length > 0
-              ? distillAnnotationsToMemory(
-                  RUNTIME_DATA_DIR,
-                  {
-                    annotations: safeCommentAttachments,
-                    userMessage: userMsg,
-                    assistantMessage: captured,
-                  },
-                  memoryOptions,
-                )
-              : Promise.resolve([]);
-          return Promise.allSettled([generalPass, annotationPass]);
-        })
-        .catch((err) => console.warn('[memory-llm] background failed', err));
-    });
-
-    // Critique Theater branch (M0 dark launch, default disabled).
-    // Only plain-stream adapters are routed through runOrchestrator in v1.
-    // Adapters that emit structured wrappers (claude-stream-json,
-    // qoder-stream-json, copilot-stream-json, json-event-stream,
-    // acp-json-rpc, pi-rpc) fall
-    // through to the legacy single-pass code path below with a one-time
-    // stderr warning so the parser never sees wrapper bytes. Per-format
-    // decoding into the orchestrator is a v2 concern.
-    //
-    // Use critiqueShouldRun (computed in the prompt builder) instead of
-    // just the env var or the rollout resolver so the orchestrator gate
-    // is in lockstep with the panel addendum. Media surfaces and runs
-    // missing brand/skill context never get the panel prompt, so they
-    // must also skip the orchestrator and fall through to legacy
-    // generation; otherwise the parser waits for <CRITIQUE_RUN> tags
-    // the model was never told to emit.
-    if (critiqueShouldRun) {
-      const adapterStreamFormat: string = def.streamFormat ?? 'plain';
-      if (adapterStreamFormat !== 'plain') {
-        if (!critiqueWarnedAdapters.has(adapterStreamFormat)) {
-          critiqueWarnedAdapters.add(adapterStreamFormat);
-          console.warn(`[critique] adapter format=${adapterStreamFormat} is not plain-stream; skipping orchestrator and falling through to legacy generation`);
-        }
-      } else {
-        const critiqueRunId = run.id;
-        // Per-run artifact directory keeps concurrent or sequential runs in the
-        // same project from overwriting each other's transcript or final HTML.
-        // Spec: artifacts/<projectId>/<runId>/transcript.ndjson(.gz).
-        const critiqueProjectKey = typeof projectId === 'string' && projectId ? projectId : critiqueRunId;
-        const critiqueArtifactDir = path.join(ARTIFACTS_DIR, critiqueProjectKey, critiqueRunId);
-        const stdoutIterable = (async function* () {
-          for await (const chunk of child.stdout) yield String(chunk);
-        })();
-        // Forward each CritiqueSseEvent on its own contract-defined channel
-        // (critique.run_started, critique.ship, critique.failed, ...) rather
-        // than wrapping the frame inside the legacy 'agent' channel. Clients
-        // that subscribe to the new event names see them directly with the
-        // contract payload as event.data.
-        //
-        // Critique events go to TWO sinks (codex P1 on PR #1338):
-        //
-        //   1. `design.runs.emit(...)` via `send(...)`, which fans out on
-        //      `/api/runs/:runId/events`. Existing transport, unchanged.
-        //   2. The per-project event-sinks map, which fans out on
-        //      `/api/projects/:projectId/events`. This is the transport the
-        //      web `CritiqueTheaterMount` actually subscribes to (the mount
-        //      is project-scoped, not run-scoped, because it lives at the
-        //      project workspace level and follows the user across runs).
-        //      Without this second sink the mount sees no frames in
-        //      production and only the e2e tests' stubbed routes deliver
-        //      anything to the reducer.
-        //
-        // The project-events route emits via `sse.send(payload.type,
-        // payload)`, so we pack the SSE channel name onto `payload.type`
-        // and let the sink push the right channel name. The web's
-        // `sseToPanelEvent` overwrites `type` from the channel name on the
-        // way back into a PanelEvent, so this round-trip stays correct.
-        const critiqueProjectIdForBus =
-          typeof projectId === 'string' && projectId ? projectId : null;
-        const critiqueBus = {
-          emit: (e) => {
-            // Two transports for every critique event: the run-scoped
-            // SSE send back to the originating chat run, plus the
-            // project-scoped fan-out so the Theater mount (subscribed
-            // to /api/projects/:id/events) sees it too. Route the
-            // project fan-out through emitProjectEvent so empty-sink
-            // cleanup and any future broadcast policy (rate limiting,
-            // schema validation, telemetry) apply uniformly across
-            // every project emitter (PerishCode P3 on PR #1338).
-            send(e.event, e.data);
-            if (critiqueProjectIdForBus) {
-              emitProjectEvent(critiqueProjectIdForBus, { ...e.data, type: e.event });
-            }
-          },
-        };
-
-        // Register this run with the in-process registry so the interrupt
-        // endpoint can cascade an AbortController to the orchestrator. The
-        // register call must run BEFORE runOrchestrator is invoked, so a
-        // request that arrives between spawn and orchestrator-start cannot
-        // miss a runId that already has a live child process.
-        const critiqueAbort = new AbortController();
-        critiqueRunRegistry.register({
-          runId: critiqueRunId,
-          projectId: critiqueProjectKey,
-          abort: critiqueAbort,
-          startedAt: Date.now(),
-        });
-
-        // Stderr forwarding and child.on('error') must be wired BEFORE the
-        // orchestrator awaits stdout. Otherwise a CLI that floods stderr can
-        // fill the OS pipe and deadlock the run until the total timeout, and
-        // an early child error fired before the orchestrator returns has no
-        // listener. Both registrations are idempotent and the run lifecycle
-        // is owned solely by the orchestrator's awaited result below.
-        child.stderr.on('data', (chunk) => {
-          noteAgentActivity();
-          emitVisibleAgentStderr(chunk);
-        });
-        child.on('error', (err) => {
-          flushVisibleAgentStderr();
-          send('error', createSseErrorPayload('AGENT_EXECUTION_FAILED', err.message));
-        });
-
-        // Wrap the child's close event so the orchestrator can race child
-        // exit against parser completion, abort, and timeouts in one awaited
-        // flow. Without this the orchestrator can't tell a non-zero exit
-        // apart from a clean ship and may misclassify failures.
-        const childExitPromise = new Promise<{ code: number | null; signal: NodeJS.Signals | null }>((resolve) => {
-          child.once('close', (code, signal) => {
-            flushVisibleAgentStderr();
-            resolve({ code, signal });
-          });
-        });
-        try {
-          const orchestratorResult = await runOrchestrator({
-            runId: critiqueRunId,
-            projectId: typeof projectId === 'string' ? projectId : '',
-            conversationId: typeof conversationId === 'string' ? conversationId : null,
-            artifactId: critiqueRunId,
-            artifactDir: critiqueArtifactDir,
-            adapter: typeof agentId === 'string' ? agentId : 'unknown',
-            // Codex P2 on PR #1485: thread the resolved skill id into the
-            // orchestrator so the Phase 12 metrics carry the real label
-            // instead of falling through to 'unknown' for every live run.
-            // `effectiveSkillId` was already computed above (line ~2951) as
-            // the request skillId with a project-row fallback; pass it
-            // through verbatim, and leave the orchestrator's own default
-            // of 'unknown' for runs that genuinely have no skill assigned.
-            skill: typeof effectiveSkillId === 'string' && effectiveSkillId
-              ? effectiveSkillId
-              : undefined,
-            cfg: critiqueCfg,
-            db,
-            bus: critiqueBus,
-            stdout: stdoutIterable,
-            child,
-            childExitPromise,
-            signal: critiqueAbort.signal,
-          });
-          // Map the critique terminal status to the chat run lifecycle.
-          // 'shipped' and 'below_threshold' both ran to a ship decision and
-          // finalize as 'succeeded'; every other status (timed_out,
-          // interrupted, degraded, failed, legacy) is a failure path so the
-          // run reflects the real outcome instead of a misleading success.
-          const succeeded = orchestratorResult.status === 'shipped'
-            || orchestratorResult.status === 'below_threshold';
-          if (run.cancelRequested) {
-            design.runs.finish(run, 'canceled', 1, null);
-          } else if (succeeded) {
-            design.runs.finish(run, 'succeeded', 0, null);
-          } else {
-            design.runs.finish(run, 'failed', 1, null);
-          }
-        } catch (err) {
-          flushVisibleAgentStderr();
-          send('error', createSseErrorPayload('AGENT_EXECUTION_FAILED', err instanceof Error ? err.message : String(err)));
-          design.runs.finish(run, 'failed', 1, null);
-        } finally {
-          critiqueRunRegistry.unregister(critiqueProjectKey, critiqueRunId);
-        }
-        return;
-      }
-    }
-
-    // Structured streams (Claude Code) go through a line-delimited JSON
-    // parser that turns stream_event objects into UI-friendly events. For
-    // plain streams (most other CLIs) we forward raw chunks unchanged so
-    // the browser can append them to the assistant's text buffer.
-    let agentStreamError = null;
-    // Holds buffered plain-text stdout chunks for agents (currently
-    // antigravity) where we need to inspect the full output at close
-    // time before deciding whether to forward it. The auth-prompt guard
-    // in the close handler suppresses the buffer when the output is an
-    // OAuth prompt; otherwise the flush below sends the chunks in order.
-    const plaintextStdoutBuffer: BufferedStdoutChunk[] = [];
-    // Arrival time of the first buffered plain-text stdout chunk
-    // (antigravity). First-token timing is stamped from this value only
-    // when the buffer is actually flushed to the client at close time. If
-    // the auth-prompt guard suppresses the buffer (the OAuth login URL is
-    // printed to stdout), no token ever reaches the user, so TTFT must not
-    // be recorded for that failure mode. See PR #3412.
-    let firstBufferedStdoutAt: number | null = null;
-    // Tracks whether any stream the run is using actually emitted user-
-    // visible content or a deliverable. Only the streams routed through
-    // `sendAgentEvent` contribute to this flag; ACP sessions and plain stdout
-    // streams are covered by their own success/failure paths and the
-    // empty-output guard below skips them via `trackingSubstantiveOutput`.
-    let agentProducedOutput = false;
-    let trackingSubstantiveOutput = false;
-    // Event types that count as "the agent actually produced a response or a
-    // deliverable." Lifecycle markers (`status`), meter readings (`usage`),
-    // reasoning deltas, and tool activity deliberately do NOT count: a run can
-    // think/read/call tools and still terminate before returning text/artifacts
-    // to the user. Treat that as empty output instead of a silent success
-    // (issues #691, #4814).
-    const SUBSTANTIVE_AGENT_EVENT_TYPES = new Set([
-      'text_delta',
-      'artifact',
-    ]);
-    // First-token timing must reflect when the user actually starts seeing
-    // model output, so only token-producing events qualify. `tool_use` is
-    // deliberately excluded: a run that opens with a Read/Glob/MCP call would
-    // otherwise stamp `firstTokenAt` before any `text_delta` streamed,
-    // making `time_to_first_token_ms` / `spawn_to_first_token_ms` under-report
-    // TTFT for tool-first runs. `thinking_delta` stays in because it is the
-    // first visible model activity the user perceives.
-    const FIRST_TOKEN_AGENT_EVENT_TYPES = new Set([
-      'text_delta',
-      'thinking_delta',
-    ]);
-    const noteFirstTokenAt = (timestamp = Date.now()) => {
-      if (run.analyticsTelemetry?.firstTokenAt) return;
-      lifecycle.mark('first_token', timestamp);
-      lifecycle.mark('first_visible_output', timestamp);
-    };
-    // Subsegment markers inside `processSpawnedAt -> firstTokenAt` (#3408 §4).
-    // `cliReadyAt` is the first well-formed adapter output and is stamped for
-    // every runtime family from its own decode choke point: first JSONL line
-    // (claude-stream-json), first decoded stream event (json-event-stream /
-    // qoder / pi-rpc), first non-empty stdout chunk (plain), or first ACP
-    // JSON-RPC message (acp-json-rpc). `sessionInitDoneAt` is only observable
-    // for ACP (the resume/`session/new` ack); for stream/plain families that
-    // gap is folded into `spawn_to_first_token_remainder_ms` rather than
-    // anchored to a fabricated marker. Both are first-write-wins like
-    // `firstTokenAt` so a later chunk cannot move an already-stamped boundary.
-    const noteCliReadyAt = (timestamp = Date.now()) => {
-      if (run.analyticsTelemetry?.cliReadyAt) return;
-      run.analyticsTelemetry = {
-        ...(run.analyticsTelemetry ?? {}),
-        cliReadyAt: timestamp,
-      };
-    };
-    const noteSessionInitDoneAt = (timestamp = Date.now()) => {
-      if (run.analyticsTelemetry?.sessionInitDoneAt) return;
-      run.analyticsTelemetry = {
-        ...(run.analyticsTelemetry ?? {}),
-        sessionInitDoneAt: timestamp,
-      };
-    };
-    const noteFirstTokenFromAgentEvent = (ev) => {
-      if (ev?.type && FIRST_TOKEN_AGENT_EVENT_TYPES.has(ev.type)) {
-        noteFirstTokenAt();
-      }
-    };
-
-    // Per-run role-marker guard for non-Claude structured streams (#3247).
-    // Claude has its own per-message guards in claude-stream.ts.
-    const runGuard = createRoleMarkerGuard('run');
-    let runWarned = false;
-    const visibleStdoutControlStripper = new TerminalControlSequenceStripper();
-    const titleMarkerStripper = createAgentTitleMarkerStripper({
-      enabled: Boolean(titleGenerationRequested),
-      emitTitle: (title) => send('agent', { type: 'conversation_title', title }),
-    });
-
-    function flushAgentTitleMarkerBuffer() {
-      const visible = titleMarkerStripper.flush();
-      if (visible) emitGuardedTextDelta(visible);
-    }
-
-    function guardTextDelta(delta) {
-      return runGuard.feedText(delta);
-    }
-
-    // Shared helper for emitting guarded text deltas across all agent
-    // stream handlers (sendAgentEvent, copilot, ACP).
-    function emitGuardedTextDelta(delta: string) {
-      const safe = guardTextDelta(delta);
-      if (safe.length > 0) {
-        send('agent', { type: 'text_delta', delta: safe });
-      }
-      if (runGuard.contaminated && !runWarned) {
-        runWarned = true;
-        const warn = runGuard.warningEvent();
-        if (warn) {
-          send('agent', warn);
-          abortForRoleMarker(warn.marker);
-        }
-      }
-    }
-
-    function emitTitleFilteredGuardedTextDelta(delta: string) {
-      const visibleDelta = titleMarkerStripper.strip(delta);
-      if (!visibleDelta) return false;
-      emitGuardedTextDelta(visibleDelta);
-      return true;
-    }
-
-    // Detection-only is necessary but not sufficient: by the time we see
-    // the role marker the model has already burned tokens, and the
-    // subprocess will keep generating downstream tokens (including
-    // `tool_use` blocks built on the fabricated context) until it exits
-    // on its own. We terminate the child immediately so:
-    //   1. Token billing stops at the detection point, not at the
-    //      model's natural completion of the contaminated response.
-    //   2. `tool_use` content blocks emitted AFTER the marker cannot
-    //      reach the daemon's tool-call dispatcher. Blocks emitted
-    //      BEFORE the marker have already been dispatched; this guard
-    //      can't help with those — they're a separate hardening.
-    //   3. The UI distinguishes "completed" from "killed by safety
-    //      guard" through a structured SSE error rather than seeing a
-    //      `fabricated_role_marker` warning followed by an eventual
-    //      normal turn-end.
-    // Idempotent — multiple guard paths (per-message Claude, run-scoped
-    // non-Claude, plain stdout) can all call it.
-    let roleMarkerAbortFired = false;
-    function abortForRoleMarker(marker: string) {
-      if (roleMarkerAbortFired) return;
-      roleMarkerAbortFired = true;
-      send(
-        'error',
-        createSseErrorPayload(
-          'ROLE_MARKER_HALLUCINATION',
-          `Run terminated: model emitted fabricated role marker (\`${marker}\`). ` +
-            'No further tokens or tool calls accepted from this turn. ' +
-            'See https://github.com/nexu-io/open-design/issues/3247.',
-          { retryable: true },
-        ),
-      );
-      // ACP sessions (Hermes, Kimi, Devin, Kiro, etc.) need explicit
-      // abort because their I/O is multiplexed and they won't
-      // necessarily exit on child SIGTERM alone.
-      if (acpSession?.abort) {
-        try {
-          acpSession.abort();
-        } catch {
-          // ignore — best-effort
-        }
-      }
-      if (child && !child.killed) design.runs.signalChild(run, 'SIGTERM');
-      scheduleForcedChildShutdown();
-    }
-
-    // Per-run tool-loop guard. Agents sometimes fixate on a failing tool call
-    // and grind through dozens of identical attempts (e.g. re-running an Edit
-    // whose `old_string` never matches, or a shell assertion against an element
-    // that does not exist). Unlike the BYOK proxy path — bounded by
-    // MAX_BYOK_TOOL_LOOPS — the autonomous chat agents had no such bound. This
-    // guard observes the normalized tool_use/tool_result events EVERY agent
-    // path emits, so one instance covers Claude, Codex/OpenCode, Copilot, ACP,
-    // … It emits a one-shot `tool_loop` warning, then (in halt mode) terminates
-    // the run at a hard ceiling. Mode via OD_TOOL_LOOP_GUARD (halt|warn|off).
-    const toolLoopGuard = createToolLoopGuard({ mode: resolveToolLoopMode() });
-    let toolLoopAbortFired = false;
-
-    // Idempotent — both agent-event paths (sendAgentEvent, the Claude
-    // stream-json callback) can route a halt verdict here.
-    function abortForToolLoop(verdict: ToolLoopVerdict) {
-      if (toolLoopAbortFired) return;
-      toolLoopAbortFired = true;
-      send(
-        'error',
-        createSseErrorPayload(
-          'TOOL_LOOP_DETECTED',
-          `Run terminated: the agent repeated a failing ${verdict.toolName} call ` +
-            `${verdict.count}× without progress (\`${verdict.signature}\`). Re-check the ` +
-            'actual target — the file, the element, the command — before retrying ' +
-            'instead of resubmitting the same turn.',
-          { retryable: true },
-        ),
-      );
-      if (acpSession?.abort) {
-        try {
-          acpSession.abort();
-        } catch {
-          // ignore — best-effort
-        }
-      }
-      // Route through signalChild (not a bare child.kill) so the halt escalates
-      // to the whole process group when one exists, matching abortForRoleMarker,
-      // cancel, and the inactivity watchdog. A bare child.kill leaves Bash/build
-      // grandchildren alive to keep mutating the workspace until the forced
-      // shutdown fires — exactly the loop class this guard is meant to stop.
-      if (child && !child.killed) design.runs.signalChild(run, 'SIGTERM');
-      scheduleForcedChildShutdown();
-    }
-
-    // Feed a normalized agent event into the loop guard and act on a verdict.
-    // Safe to call for every event; non-tool events are ignored. Emit the
-    // `tool_loop` warning to the UI/CLI, and on a halt verdict tear the run
-    // down so it cannot keep grinding.
-    function observeToolEventForLoop(ev: any) {
-      if (!ev || typeof ev !== 'object') return;
-      if (ev.type === 'tool_use' && typeof ev.id === 'string') {
-        toolLoopGuard.observeToolUse(ev.id, typeof ev.name === 'string' ? ev.name : 'tool', ev.input);
-        return;
-      }
-      if (ev.type === 'tool_result' && typeof ev.toolUseId === 'string') {
-        const verdict = toolLoopGuard.observeToolResult(
-          ev.toolUseId,
-          Boolean(ev.isError),
-          typeof ev.content === 'string' ? ev.content : '',
-        );
-        if (verdict) {
-          send('agent', verdict);
-          if (verdict.action === 'halt') abortForToolLoop(verdict);
-        }
-      }
-    }
-
-    // Single choke point for emitting an agent event to the client. EVERY
-    // stream handler (sendAgentEvent, the Claude callback, Copilot, ACP, …)
-    // emits through here, never via a bare send('agent', …), so the tool-loop
-    // guard sees every runtime's tool activity and no handler can drift out of
-    // coverage. observe runs AFTER the send so a `tool_loop` warning/halt
-    // follows the result that triggered it in the stream. (PR #3375 review:
-    // Copilot and ACP bypassed the guard by calling send('agent', …) directly.)
-    function emitAgentEvent(ev: any) {
-      send('agent', ev);
-      observeToolEventForLoop(ev);
-    }
-
-    const sendAgentEvent = (ev) => {
-      if (ev?.type === 'error') {
-        if (agentStreamError) return;
-        flushVisibleAgentStderr();
-        const failureText = [
-          String(ev.message || 'Agent stream error'),
-          typeof ev.raw === 'string' ? ev.raw : '',
-          agentStdoutTail,
-          agentStderrTail,
-        ].join('\n');
-        agentStreamError = rewriteKnownAgentStreamError(
-          agentId,
-          String(ev.message || 'Agent stream error'),
-          failureText,
-        );
-        clearInactivityWatchdog();
-        const authFailure = classifyAgentAuthFailure(agentId, failureText);
-        if (authFailure?.status === 'missing') {
-          send('error', createSseErrorPayload(
-            'AGENT_AUTH_REQUIRED',
-            authFailure.message ?? cursorAuthGuidance(),
-            { retryable: true },
-          ));
-          return;
-        }
-        // Recover the specific model-service failure class (auth / quota /
-        // upstream) for agents without a tailored probe (Claude Code, codex,
-        // …), so the chat shows an accurate reason instead of the generic
-        // execution-failed bucket.
-        const serviceCode = classifyAgentServiceFailure(failureText);
-        if (serviceCode) {
-          send('error', createSseErrorPayload(serviceCode, agentStreamError, {
-            details: ev.raw ? { raw: ev.raw } : undefined,
-            retryable: true,
-          }));
-          return;
-        }
-        send('error', createSseErrorPayload('AGENT_EXECUTION_FAILED', agentStreamError, {
-          details: ev.raw ? { raw: ev.raw } : undefined,
-          retryable: false,
-        }));
-        return;
-      }
-      // First well-formed decoded stream event = CLI ready for the
-      // json-event-stream / qoder / pi-rpc families (#3408 §4 marker).
-      noteCliReadyAt();
-      // Capture-style resume: codex reports its own thread id on the
-      // `thread.started` status event. Persist the most recent non-empty id we
-      // see so the create-turn store (and the resumable-failure store) use the
-      // CLI's real session handle, not the unused daemon-minted `newSessionId`.
-      if (
-        agentCapturesSessionId &&
-        ev?.type === 'status' &&
-        typeof ev.sessionId === 'string' &&
-        ev.sessionId.length > 0
-      ) {
-        capturedSessionId = ev.sessionId;
-      }
-      lastAgentEventPhase = summarizeAgentEventForInactivity(ev);
-      noteAgentActivity();
-      // Role-marker guard for qoder / json-event-stream / pi-rpc (#3247).
-      if (ev?.type === 'text_delta' && typeof ev.delta === 'string') {
-        if (emitTitleFilteredGuardedTextDelta(ev.delta)) {
-          noteFirstTokenAt();
-          agentProducedOutput = true;
-        }
-        return;
-      }
-      noteFirstTokenFromAgentEvent(ev);
-      if (ev?.type && SUBSTANTIVE_AGENT_EVENT_TYPES.has(ev.type)) {
-        agentProducedOutput = true;
-      }
-      emitAgentEvent(ev);
-    };
-    const parseBufferedAntigravityGeminiJsonEventStream = () => {
-      if (
-        def.id !== 'antigravity' ||
-        plaintextStdoutBuffer.length === 0
-      ) {
-        return false;
-      }
-      const bufferedStdout = plaintextStdoutBuffer.map((chunk) => chunk.text).join('');
-      if (!looksLikeGeminiJsonEventStream(bufferedStdout)) return false;
-      trackingSubstantiveOutput = true;
-      const firstTokenAt = bufferedAntigravityGeminiFirstTokenAt(plaintextStdoutBuffer);
-      if (firstTokenAt !== null) noteFirstTokenAt(firstTokenAt);
-      const handler = createJsonEventStreamHandler('gemini', sendAgentEvent);
-      handler.feed(bufferedStdout);
-      handler.flush();
-      plaintextStdoutBuffer.length = 0;
-      return true;
-    };
-
-    if (def.streamFormat === 'claude-stream-json') {
-      const claude = createClaudeStreamHandler((ev) => {
-        // First parsed claude-stream-json event = CLI ready (#3408 §4); the
-        // init/system line arrives well before the model's first token.
-        noteCliReadyAt();
-        if (ev?.type === 'error') {
-          if (agentStreamError) return;
-          flushVisibleAgentStderr();
-          const message = String((ev as any).message || 'Claude Code stream error');
-          const failureText = [
-            message,
-            typeof (ev as any).code === 'string' ? (ev as any).code : '',
-            agentStdoutTail,
-            agentStderrTail,
-          ].join('\n');
-          clearInactivityWatchdog();
-          // Claude surfaces a connection drop / reset as an in-stream `error`
-          // frame (assistant `error:"unknown"` + the raw SDK string), which
-          // would otherwise reach the UI verbatim as a non-retryable
-          // AGENT_EXECUTION_FAILED. Run the same per-agent diagnostic used at
-          // child-exit so this path emits the specific class
-          // (AGENT_CONNECTION_DROPPED) — retryable, with copy the web can
-          // localize and triage can count by code.
-          const diagnostic = diagnoseClaudeCliFailure({
-            agentId: def.id,
-            exitCode: 1,
-            stderrTail: agentStderrTail,
-            stdoutTail: failureText,
-            env: spawnedAgentEnv,
-            resolvedBin: agentLaunch.selectedPath,
-          });
-          const serviceCode = classifyAgentServiceFailure(failureText);
-          agentStreamError = diagnostic?.message
-            ?? rewriteKnownAgentStreamError(agentId, message, failureText);
-          send('error', createSseErrorPayload(
-            diagnostic?.code ?? serviceCode ?? 'AGENT_EXECUTION_FAILED',
-            agentStreamError,
-            {
-              retryable: diagnostic?.retryable
-                ?? (serviceCode === 'AGENT_AUTH_REQUIRED' || serviceCode === 'RATE_LIMITED'),
-              ...(diagnostic ? { details: { detail: diagnostic.detail } } : {}),
-            },
-          ));
-          return;
-        }
-        lastAgentEventPhase = summarizeAgentEventForInactivity(ev);
-        noteAgentActivity();
-        if (ev?.type === 'text_delta' && typeof ev.delta === 'string') {
-          const visibleDelta = titleMarkerStripper.strip(ev.delta);
-          if (visibleDelta) {
-            noteFirstTokenAt();
-            emitAgentEvent({ ...ev, delta: visibleDelta });
-          }
-          return;
-        }
-        noteFirstTokenFromAgentEvent(ev);
-        emitAgentEvent(ev);
-        // Claude uses per-message guards (claude-stream.ts) rather than the
-        // run-scoped guard above, so its `fabricated_role_marker` events
-        // surface here directly from the stream handler, not via
-        // emitGuardedTextDelta. Same abort semantics apply.
-        if (ev && (ev as any).type === 'fabricated_role_marker') {
-          const m = (ev as any).marker;
-          abortForRoleMarker(typeof m === 'string' ? m : 'role marker');
-        }
-        // Stream-json input mode keeps the child's stdin open across the
-        // turn so the daemon can stream further user messages mid-turn. The
-        // child has no other way to know the turn is over, though — without
-        // an EOF it sits idle until the inactivity watchdog kills it.
-        // Bookkeeping here closes stdin on a clean terminal turn:
-        //   - turn_end (per-turn synthesized from `stop_reason`): fire on
-        //     `end_turn` etc. but NOT on `tool_use` — that stop reason
-        //     means the model paused mid-tool, not "turn complete".
-        //   - usage (session result at EOF in single-shot mode).
-        try {
-          applyClaudeStreamJsonRunBookkeeping(run, ev);
-        } catch {}
-      }, { suppressHtmlArtifactsAfterFileWrite: def.id === 'claude' });
-      child.stdout.on('data', (chunk) => claude.feed(chunk));
-      child.on('close', () => claude.flush());
-    } else if (def.streamFormat === 'qoder-stream-json') {
-      trackingSubstantiveOutput = true;
-      const qoder = createQoderStreamHandler(sendAgentEvent);
-      child.stdout.on('data', (chunk) => qoder.feed(chunk));
-      child.on('close', () => qoder.flush());
-    } else if (def.streamFormat === 'copilot-stream-json') {
-      const copilot = createCopilotStreamHandler((ev) => {
-        lastAgentEventPhase = summarizeAgentEventForInactivity(ev);
-        noteAgentActivity();
-        if (ev?.type === 'text_delta' && typeof ev.delta === 'string') {
-          if (emitTitleFilteredGuardedTextDelta(ev.delta)) {
-            noteFirstTokenAt();
-          }
-          return;
-        }
-        noteFirstTokenFromAgentEvent(ev);
-        emitAgentEvent(ev);
-      });
-      child.stdout.on('data', (chunk) => copilot.feed(chunk));
-      child.on('close', () => copilot.flush());
-    } else if (def.streamFormat === 'pi-rpc') {
-      // Route through sendAgentEvent so that pi-rpc's error events
-      // (extension_error, auto_retry_end with success=false, and the
-      // message_update error delta) set agentStreamError and flip the
-      // run to `failed` on close — same path as qoder-stream-json and
-      // json-event-stream after issue #691. Also enables the
-      // substantive-output guard (agentProducedOutput) so a pi run
-      // that exits 0 without producing visible content is caught.
-      //
-      // attachPiRpcSession invokes its send callback with the two-arg
-      // channel/payload shape: send('agent', payload) for normal events
-      // and send('error', {message}) from fail(). sendAgentEvent
-      // expects a single event object, so we adapt at the call site:
-      //   - 'agent' channel → relay payload through sendAgentEvent
-      //   - 'error' channel → route through the daemon's error path
-      //     (createSseErrorPayload + send SSE + set agentStreamError)
-      trackingSubstantiveOutput = true;
-      acpSession = attachPiRpcSession({
-        child,
-        prompt: composed,
-        cwd: effectiveCwd,
-        model: safeModel,
-        parentSession: agentResumeCtx.isResuming && agentResumeCtx.resumeSessionId
-          ? agentResumeCtx.resumeSessionId
-          : undefined,
-        send: (channel, payload) => {
-          if (channel === 'agent') {
-            sendAgentEvent(payload);
-          } else if (channel === 'error') {
-            if (agentStreamError) return;
-            flushVisibleAgentStderr();
-            agentStreamError = String(payload?.message || 'Pi session error');
-            const piErrorCode = typeof payload?.code === 'string' ? payload.code : null;
-            if (piErrorCode) {
-              run.errorCode = piErrorCode;
-            }
-            if (piErrorCode === 'PI_PARENT_SESSION_FAILED' && run.conversationId) {
-              clearAgentSession(db, run.conversationId, def.id);
-            }
-            clearInactivityWatchdog();
-            send('error', createSseErrorPayload(
-              'AGENT_EXECUTION_FAILED',
-              agentStreamError,
-              { retryable: false },
-            ));
-          } else {
-            noteAgentActivity();
-            send(channel, payload);
-          }
-        },
-        imagePaths: def.supportsImagePaths ? amrStagedImages : [],
-        uploadRoot: UPLOAD_DIR,
-      });
-    } else if (def.streamFormat === 'acp-json-rpc') {
-      const acpStageTimeoutMs = resolveAcpStageTimeoutMs(def.inactivityTimeoutMs);
-      acpSession = attachAcpSession({
-        child,
-        prompt: composed,
-        cwd: effectiveCwd,
-        model: safeModel,
-        imagePaths: def.supportsImagePaths ? amrStagedImages : [],
-        mcpServers,
-        envFormat: def.acpMcpEnvFormat ?? 'array',
-        executionProfile,
-        ...(def.id === 'amr' ? { modelUnavailableErrorCode: 'AMR_MODEL_UNAVAILABLE' } : {}),
-        // Resume the prior upstream session (drives `session/load`) when the
-        // resume-identity guard says it is safe; otherwise a fresh session/new.
-        ...(def.resumesSessionViaAcpLoad === true && agentResumeCtx.isResuming && agentResumeCtx.resumeSessionId
-          ? { resumeSessionId: agentResumeCtx.resumeSessionId }
-          : {}),
-        onCliReady: () => noteCliReadyAt(),
-        onSessionInit: () => noteSessionInitDoneAt(),
-        send: (event, data) => {
-          if (event === 'agent') {
-            lastAgentEventPhase = summarizeAgentEventForInactivity(data);
-          }
-          noteAgentActivity();
-          if (event === 'error') flushVisibleAgentStderr();
-          if (def.id === 'amr' && event === 'error') {
-            const failure = classifyAmrAccountFailureSignal({
-              details: data?.error?.details,
-              message: data?.message,
-              errorMessage: data?.error?.message,
-              errorCode: data?.error?.code,
-              stdoutTail: agentStdoutTail,
-              stderrTail: agentStderrTail,
-            });
-            if (failure) {
-              sendAmrAccountFailure(failure);
-              return;
-            }
-          }
-          // Hold back the `resume_failed` error so the same-turn reseed stays
-          // transparent. When this run is resuming an upstream session via
-          // `session/load` and the agent reports that session is gone, the ACP
-          // bridge has already called `fail()` -> `send('error')` for the failed
-          // load. The child-close handler then clears the stale handle and
-          // re-runs this turn fresh (the resume-target-missing block below), so
-          // forwarding this error would flash an execution failure — and trip
-          // clients that treat an SSE `error` as terminal — a beat before the
-          // invisible recovery. Suppress it and leave a diagnostic instead; the
-          // close handler is the sole authority on whether this turn ends in an
-          // error or a transparent reseed. The `resumeAutoReseeded` guard lets a
-          // second resume failure in one run fall through to the explicit
-          // "resend your message" affordance the close handler emits.
-          if (
-            event === 'error' &&
-            def.resumesSessionViaAcpLoad === true &&
-            agentResumeCtx.isResuming &&
-            agentResumeCtx.resumeSessionId &&
-            !run.resumeAutoReseeded &&
-            isAgentResumeFailure(def.id, agentStderrTail, agentStdoutTail)
-          ) {
-            design.runs.emit(run, 'diagnostic', {
-              type: 'agent_resume_failed_suppressed',
-              agent_id: def.id,
-              reason: 'resume_failed',
-              previous_session_id: agentResumeCtx.resumeSessionId ?? null,
-            });
-            return;
-          }
-          if (event === 'agent' && data?.type === 'text_delta' && typeof data.delta === 'string') {
-            if (emitTitleFilteredGuardedTextDelta(data.delta)) {
-              noteFirstTokenAt();
-            }
-            return;
-          }
-          if (event === 'agent') {
-            noteFirstTokenFromAgentEvent(data);
-            emitAgentEvent(data);
-          } else {
-            send(event, data);
-          }
-        },
-        ...(acpStageTimeoutMs !== undefined ? { stageTimeoutMs: acpStageTimeoutMs } : {}),
-      });
-    } else if (def.streamFormat === 'json-event-stream') {
-      // Pipe through sendAgentEvent so the OpenCode `type:'error'` frame
-      // (now emitted as a real error event by json-event-stream.ts after
-      // #691) actually triggers `agentStreamError` instead of being
-      // forwarded as a no-op `agent` SSE event. This also wires the
-      // substantive-output tracking the close handler reads below.
-      trackingSubstantiveOutput = true;
-      const handler = createJsonEventStreamHandler(
-        def.eventParser || def.id,
-        sendAgentEvent,
-      );
-      child.stdout.on('data', (chunk) => handler.feed(chunk));
-      child.on('close', () => handler.flush());
-    } else if (def.id === 'antigravity') {
-      // Buffer stdout until close so the auth-prompt guard can suppress
-      // the OAuth URL before forwarding it to the client as assistant
-      // text. agy exits 0 after printing the auth URL on stdout, so the
-      // chunks would otherwise arrive before the close-time classifier
-      // detects them as an auth prompt. First-token timing is deliberately
-      // NOT stamped here — only the first chunk's arrival time is recorded,
-      // and `firstTokenAt` is stamped from it at flush time so the
-      // suppressed OAuth-prompt path never reports a TTFT (PR #3412).
-      child.stdout.on('data', (chunk) => {
-        noteAgentActivity();
-        const receivedAt = Date.now();
-        if (firstBufferedStdoutAt === null) firstBufferedStdoutAt = receivedAt;
-        plaintextStdoutBuffer.push({ text: String(chunk), receivedAt });
-      });
-    } else {
-      // Plain / BYOK mode: guard raw stdout chunks (#3247).
-      child.stdout.on('data', (chunk) => {
-        noteAgentActivity();
-        const text = typeof chunk === 'string' ? chunk : String(chunk);
-        // First non-empty stdout chunk = CLI ready for the plain family
-        // (#3408 §4 marker). A plain adapter has no structured preamble, so
-        // this typically coincides with its first model output.
-        if (text.length > 0) noteCliReadyAt();
-        const strippedText = visibleStdoutControlStripper.write(text);
-        const visibleText = titleMarkerStripper.strip(strippedText);
-        const safe = guardTextDelta(visibleText);
-        if (safe.length > 0) {
-          noteFirstTokenAt();
-          send('stdout', { chunk: safe });
-        }
-        if (runGuard.contaminated && !runWarned) {
-          runWarned = true;
-          const warn = runGuard.warningEvent();
-          if (warn) {
-            send('agent', warn);
-            abortForRoleMarker(warn.marker);
-          }
-        }
-      });
-    }
-    // Wire the acpSession onto the run so cancel() can call abort()
-    // instead of raw SIGTERM (applies to pi-rpc and acp-json-rpc).
-    run.acpSession = acpSession;
-    child.stderr.on('data', (chunk) => {
-      noteAgentActivity();
-      emitVisibleAgentStderr(chunk);
-    });
-
-    child.on('error', (err) => {
-      clearInactivityWatchdog();
-      cleanupPromptFile();
-      flushVisibleAgentStderr();
-      revokeToolToken('child_exit');
-      unregisterChatAgentEventSink();
-      send('error', createSseErrorPayload('AGENT_EXECUTION_FAILED', err.message));
-      finishWithRetryDecision('failed', 1, null);
-    });
-    child.on('close', async (code, signal) => {
-      try {
-      clearInactivityWatchdog();
-      clearForcedChildShutdown();
-      flushVisibleAgentStderr();
-      if (watchdogRetryRestarted) {
-        // The inactivity watchdog already failed this attempt and the same-run
-        // retry restarted on a fresh child. Finalization and event-sink / run-
-        // handle ownership (keyed by the shared runId) now belong to the new
-        // attempt, so this stalled child's close must not re-run them — doing
-        // so would re-finalize the run and delete the new attempt's sink.
-        // Revoke only THIS attempt's tool token (idempotent, keyed by its own
-        // token string) and bail; the `finally` block still cleans up logs.
-        revokeToolToken('child_exit');
-        return;
-      }
-      revokeToolToken('child_exit');
-      unregisterChatAgentEventSink();
-      // Resume-target-missing recovery runs BEFORE the generic fatal/stream-error
-      // short-circuits. The signal arrives differently per adapter: codex reports
-      // "no rollout found for thread id" as a stream `error` event, while AMR/vela
-      // reports a structured `resume_failed` JSON-RPC error that the ACP bridge
-      // turns into a FATAL. Either would otherwise be swallowed by the
-      // `fatal_rpc_error` / `stream_error` paths below and leave the dead session
-      // id stored — so every later turn would retry the same broken resume (#4275
-      // class). Clearing the stale handle here lets the next turn start fresh +
-      // re-seed the full transcript: one cold turn, never a broken conversation.
-      if (
-        !run.cancelRequested &&
-        (def.resumesSessionViaCli === true || def.resumesSessionViaAcpLoad === true) &&
-        agentResumeCtx.isResuming &&
-        run.conversationId &&
-        isAgentResumeFailure(def.id, agentStderrTail, agentStdoutTail)
-      ) {
-        // The resumed upstream session is gone (expired / pruned). Clear the dead
-        // handle and TRANSPARENTLY re-run this same turn with a fresh session +
-        // the full transcript rebuilt from the DB — exactly the pre-session-reuse
-        // path. The user sees one (slightly slower) turn, never an error or a
-        // "resend" prompt. Re-spawn reuses the same-run retry machinery; because
-        // the session row is now cleared, the re-spawn resolves isResuming=false
-        // (fresh session, full transcript), so it CANNOT resume-fail again — the
-        // `resumeAutoReseeded` guard is belt-and-suspenders against any loop.
-        clearAgentSession(db, run.conversationId, def.id);
-        if (!run.resumeAutoReseeded) {
-          run.resumeAutoReseeded = true;
-          run.resumeAutoReseededFrom = agentResumeCtx.resumeSessionId ?? null;
-          // Persisted to the per-run events.jsonl that the help → diagnostics
-          // export bundles, so the whole resume → fail → auto-reseed chain is
-          // visible in a support bundle without any user-facing signal.
-          design.runs.emit(run, 'diagnostic', {
-            type: 'agent_resume_auto_reseed',
-            agent_id: def.id,
-            reason: 'resume_failed',
-            previous_session_id: agentResumeCtx.resumeSessionId ?? null,
-            stale_session_cleared: true,
-          });
-          scheduleRetryRestart(0);
-          return;
-        }
-        // Unreachable in practice (the reseed runs fresh); if a second resume
-        // failure ever surfaces in one run, fall back to the explicit affordance.
-        send('error', createSseErrorPayload(
-          'AGENT_EXECUTION_FAILED',
-          'The previous session could not be resumed (it may have expired). Resend your message to continue with a fresh session.',
-          { retryable: true },
-        ));
-        return design.runs.finish(run, 'failed', code ?? 1, signal ?? null);
-      }
-      if (acpSession?.hasFatalError()) {
-        markRpcCloseReason('fatal_rpc_error');
-        return finishWithRetryDecision('failed', code ?? 1, signal ?? null);
-      }
-      parseBufferedAntigravityGeminiJsonEventStream();
-      flushAgentTitleMarkerBuffer();
-      if (agentStreamError) {
-        markRpcCloseReason('stream_error');
-        return finishWithRetryDecision('failed', code === 0 ? 1 : (code ?? 1), signal ?? null);
-      }
-      if (
-        code !== 0 &&
-        !run.cancelRequested
-      ) {
-        if (def.id === 'amr') {
-          const amrFailure = classifyAmrAccountFailureSignal({
-            stdoutTail: agentStdoutTail,
-            stderrTail: agentStderrTail,
-          });
-          if (amrFailure) {
-            sendAmrAccountFailure(amrFailure);
-            return finishWithRetryDecision('failed', code ?? 1, signal ?? null);
-          }
-        }
-        const authFailure = classifyAgentAuthFailure(
-          agentId,
-          `${agentStderrTail}\n${agentStdoutTail}`,
-        );
-        if (authFailure?.status === 'missing') {
-          send('error', createSseErrorPayload(
-            'AGENT_AUTH_REQUIRED',
-            authFailure.message ?? cursorAuthGuidance(),
-            { retryable: true },
-          ));
-          return finishWithRetryDecision('failed', code ?? 1, signal ?? null);
-        }
-      }
-      // Empty-output guard: a clean `code === 0` exit with no visible
-      // output means the run silently finished without producing anything.
-      // Surface an explicit failure so the chat shows a clear reason.
-      if (
-        code === 0 &&
-        !run.cancelRequested &&
-        trackingSubstantiveOutput &&
-        !agentProducedOutput
-      ) {
-        markRpcCloseReason('empty_output');
-        send('error', createSseErrorPayload(
-          'AGENT_EXECUTION_FAILED',
-          'Agent completed without producing any output. The model or provider may have returned an empty response — check the agent logs for upstream errors.',
-          { retryable: true },
-        ));
-        return finishWithRetryDecision('failed', code, signal);
-      }
-      if (
-        code === 0 &&
-        !run.cancelRequested &&
-        isPluginAuthoringRun(db, run, getSnapshot) &&
-        !(await hasGeneratedPluginArtifacts(cwd)) &&
-        !emittedRenderableQuestionForm(clarifyingQuestionText)
-      ) {
-        send('error', createSseErrorPayload(
-          'AGENT_EXECUTION_FAILED',
-          'Plugin authoring ended before generating the required generated-plugin artifacts.',
-          { retryable: true },
-        ));
-        return finishWithRetryDecision('failed', code, signal);
-      }
-      // Plain-stream auth-failure guard: plain adapters (today
-      // antigravity, deepseek's TUI variants) may exit cleanly with
-      // visible stdout that's actually an auth prompt — agy prints
-      // "Authentication required. Please visit the URL to log in:
-      // <URL>" + "Error: authentication timed out." rather than
-      // failing with a non-zero exit. Without this guard the chat
-      // shows that raw prompt as the agent's "reply", and the user
-      // has no way to actually complete OAuth from inside the chat.
-      // Override the apparent success with a proper
-      // AGENT_AUTH_REQUIRED error carrying actionable guidance.
-      if (
-        code === 0 &&
-        !run.cancelRequested &&
-        !trackingSubstantiveOutput &&
-        childStdoutSeen
-      ) {
-        const authFailure = classifyAgentAuthFailure(
-          agentId,
-          `${agentStderrTail}\n${agentStdoutTail}`,
-        );
-        if (authFailure?.status === 'missing') {
-          send('error', createSseErrorPayload(
-            'AGENT_AUTH_REQUIRED',
-            authFailure.message ?? `${def.name} authentication required. Please re-authenticate and retry.`,
-            { retryable: true },
-          ));
-          return finishWithRetryDecision('failed', 0, signal);
-        }
-      }
-      // Plain-stream empty-output guard: plain agents send raw stdout
-      // chunks without structured event tracking. Detect auth failures
-      // and quota / upstream errors when exit 0 but no stdout was
-      // seen. agy in print mode is silent on stdout/stderr for both
-      // missing-auth AND quota-exhausted failures; the daemon piped
-      // agy's `--log-file` to `agentLogFilePath` precisely so this
-      // guard can grep the upstream error code (RESOURCE_EXHAUSTED 429
-      // for quota, "not logged into Antigravity" for auth) and route
-      // to the right user-facing guidance.
-      if (
-        code === 0 &&
-        !run.cancelRequested &&
-        !trackingSubstantiveOutput &&
-        !childStdoutSeen
-      ) {
-        markRpcCloseReason('empty_output');
-        let combinedDetail = `${agentStderrTail}\n${agentStdoutTail}`;
-        if (def.id === 'antigravity' && agentLogFilePath) {
-          try {
-            const logContent = await fs.promises.readFile(agentLogFilePath, 'utf8');
-            // Keep the last 8 KB — quota / auth lines all land near the
-            // tail (after the spawn / model-config preamble).
-            combinedDetail = `${combinedDetail}\n${logContent.slice(-8192)}`;
-          } catch {
-            // Missing log file (agy didn't write it, mounted tmpfs is
-            // read-only, etc.) is fine — fall through to the generic
-            // empty-output message.
-          }
-        }
-        const authFailure = classifyAgentAuthFailure(agentId, combinedDetail);
-        const serviceFailure = !authFailure
-          ? classifyAgentServiceFailure(combinedDetail)
-          : null;
-        const isAntigravityQuota =
-          def.id === 'antigravity' && serviceFailure === 'RATE_LIMITED';
-        // Antigravity-only fallback: if neither classifier matched but
-        // the run was silent, lean on the empirical observation that
-        // an empty agy print-mode exit almost always means
-        // missing-OAuth (the only other silent path is quota, which
-        // the log-file check above already caught).
-        const useAntigravityAuthFallback =
-          !authFailure && !serviceFailure && def.id === 'antigravity';
-        const errorCode =
-          authFailure || useAntigravityAuthFallback
-            ? 'AGENT_AUTH_REQUIRED'
-            : isAntigravityQuota
-              ? 'RATE_LIMITED'
-              : 'AGENT_EXECUTION_FAILED';
-        const msg = authFailure
-          ? authFailure.message ?? `${def.name} authentication expired. Please re-authenticate and retry.`
-          : isAntigravityQuota
-            ? antigravityQuotaGuidance()
-            : useAntigravityAuthFallback
-              ? antigravityAuthGuidance()
-              : `${def.name} returned an empty response. This may indicate an expired session — try re-authenticating the agent.`;
-        send('error', createSseErrorPayload(
-          errorCode,
-          msg,
-          { retryable: true },
-        ));
-        return finishWithRetryDecision('failed', 0, signal);
-      }
-      // ACP agents that don't shut down on stdin.end() (e.g. Devin for
-      // Terminal) are forced to exit via SIGTERM from attachAcpSession after
-      // a clean prompt completion. Without an override, the chat run would
-      // be marked `failed` because `code === 0` fails (code is null on a
-      // signal exit). `completedSuccessfully()` reports whether the ACP
-      // session resolved without a fatal error or abort.
-      //
-      // Scope the override narrowly to the exact forced-shutdown shape this
-      // PR introduces: code is null AND signal is SIGTERM AND the ACP
-      // session reported clean completion. Any other post-response failure
-      // (non-zero exit code, SIGKILL, SIGSEGV, etc.) still propagates as
-      // `failed`, preserving the existing close-status behavior for genuine
-      // post-response process problems.
-      const acpCleanCompletion =
-        typeof acpSession?.completedSuccessfully === 'function' &&
-        acpSession.completedSuccessfully();
-      const runArtifactSideEffects = scanRunEventsForRetrySideEffects(run.events);
-      const status = classifyChatRunCloseStatus({
-        cancelRequested: !!run.cancelRequested,
-        code,
-        signal,
-        acpCleanCompletion,
-        artifactQuietShutdownRequested,
-        turnCompletedCleanly: !!run.turnCompletedCleanly,
-        artifactProducedThisRun:
-          runArtifactSideEffects.artifactWriteSeen ||
-          runArtifactSideEffects.liveArtifactSeen,
-      });
-      // Skip the close-handler failure emit when the run is already
-      // terminal: the inactivity watchdog (failForInactivity) finishes the
-      // run — sending its error and clearing run.clients/eventsLogStream —
-      // before SIGTERM, so re-emitting here would double-send the error and
-      // reopen the closed events-log stream. The run is finalized below
-      // regardless (finish() no-ops once terminal).
-      if (status === 'failed' && !design.runs.isTerminal(run.status)) {
-        const diagnostic = diagnoseClaudeCliFailure({
-          agentId: def.id,
-          exitCode: code,
-          signal,
-          stderrTail: agentStderrTail,
-          stdoutTail: agentStdoutTail,
-          env: spawnedAgentEnv,
-          resolvedBin: agentLaunch.selectedPath,
-        });
-        // A non-zero exit whose output reads as an auth / quota / upstream
-        // problem (typical of Claude Code, codex, …) gets the specific code
-        // rather than the generic execution-failed bucket; the human-readable
-        // message still prefers the richer CLI diagnostic when we have one.
-        const serviceCode = classifyAgentServiceFailure(
-          `${agentStderrTail}\n${agentStdoutTail}`,
-        );
-        if (diagnostic) {
-          send('error', createSseErrorPayload(
-            // A diagnostic that named its own failure class (e.g.
-            // AGENT_CONNECTION_DROPPED) wins over the generic service-failure
-            // sniff so the UI can localize by code and triage can count it.
-            diagnostic.code ?? serviceCode ?? 'AGENT_EXECUTION_FAILED',
-            diagnostic.message,
-            { retryable: diagnostic.retryable, details: { detail: diagnostic.detail } },
-          ));
-        } else if (serviceCode) {
-          const detail = (agentStderrTail || agentStdoutTail || '').trim();
-          send('error', createSseErrorPayload(
-            serviceCode,
-            detail || 'The model service returned an error.',
-            { retryable: true },
-          ));
-        } else {
-          // OpenCode swallows provider failures in headless mode: a 429
-          // usage-limit is marked retryable and retried silently with
-          // nothing on stdout/stderr, so the run only dies via the
-          // inactivity watchdog and the checks above find no signal. The
-          // real reason is recorded only in OpenCode's own session log,
-          // so recover it before falling back to the generic rewrite.
-          // See issue #982.
-          const openCodeFailure =
-            def.id === 'opencode'
-              ? readOpenCodeServiceFailure(spawnedAgentEnv, { since: run.createdAt })
-              : null;
-          if (openCodeFailure) {
-            send('error', createSseErrorPayload(
-              openCodeFailure.code,
-              openCodeFailure.message,
-              { retryable: openCodeFailure.retryable },
-            ));
-          } else {
-            const rewritten = rewriteKnownAgentStreamError(
-              def.id,
-              (agentStderrTail || agentStdoutTail || '').trim(),
-              `${agentStderrTail}\n${agentStdoutTail}`,
-            );
-            if (rewritten !== 'Agent stream error') {
-              send('error', createSseErrorPayload(
-                'AGENT_EXECUTION_FAILED',
-                rewritten,
-                { retryable: true },
-              ));
-            }
-          }
-        }
-      }
-      // Reconcile any HTML artifacts that were written during this run
-      // without a manifest sidecar (e.g. agent used write_file instead of
-      // create_artifact, or the run terminated between HTML write and
-      // sidecar write). Only files modified after the run started are
-      // touched — pre-existing HTML in imported-folder projects must not
-      // receive spurious manifests. Best-effort; must not block finalisation.
-      // See issue #2893.
-      if (run.projectId) {
-        (async () => {
-          try {
-            const project = getProject(db, run.projectId);
-            const files = await listFiles(PROJECTS_DIR, run.projectId, {
-              metadata: project?.metadata,
-            });
-            const dir = resolveProjectDir(PROJECTS_DIR, run.projectId, project?.metadata);
-            for (const f of files) {
-              const ext = f.name.slice(f.name.lastIndexOf('.')).toLowerCase();
-              if (ext !== '.html' && ext !== '.htm') continue;
-              try {
-                const filePath = path.join(dir, f.name);
-                const st = await fs.promises.stat(filePath);
-                if (!isRunTouchedProjectFile(st.mtimeMs, runStartTimeMs)) continue;
-                await reconcileHtmlArtifactManifest(
-                  PROJECTS_DIR,
-                  run.projectId,
-                  f.name,
-                  project?.metadata,
-                );
-              } catch { /* per-file best-effort */ }
-            }
-          } catch { /* project-level best-effort */ }
-        })();
-      }
-      // Flush buffered plain-text stdout (antigravity) that was not
-      // suppressed by the auth-prompt guard above. Send each chunk in
-      // order before finishing so the assistant text arrives before the
-      // run's `finished` event. Stamp first-token timing here — and only
-      // here — using the first chunk's arrival time, so the OAuth-prompt
-      // path (which returns before this flush) never records a TTFT for
-      // output the user never saw (PR #3412).
-      if (plaintextStdoutBuffer.length > 0 && firstBufferedStdoutAt !== null) {
-        noteFirstTokenAt(firstBufferedStdoutAt);
-      }
-      for (const chunk of plaintextStdoutBuffer) {
-        const strippedText = visibleStdoutControlStripper.write(chunk.text);
-        const visibleText = titleMarkerStripper.strip(strippedText);
-        if (visibleText) send('stdout', { chunk: visibleText });
-      }
-      const flushedControlText = visibleStdoutControlStripper.flush();
-      const flushedTitleMarkerText =
-        titleMarkerStripper.strip(flushedControlText) + titleMarkerStripper.flush();
-      if (flushedTitleMarkerText) send('stdout', { chunk: flushedTitleMarkerText });
-      if (
-        status === 'succeeded' &&
-        (def.streamFormat ?? 'plain') === 'plain' &&
-        run.projectId
-      ) {
-        const plainStdout = plainStdoutFromRunEvents(run.events);
-        if (plainStdout.includes('<artifact')) {
-          try {
-            const project = getProject(db, run.projectId);
-            const persistedPlainArtifacts = await persistPlainStreamArtifacts({
-              projectsRoot: PROJECTS_DIR,
-              projectId: run.projectId,
-              stdout: plainStdout,
-              metadata: project?.metadata,
-              writeProjectFile,
-            });
-            if (persistedPlainArtifacts.length > 0) {
-              for (const artifact of persistedPlainArtifacts) {
-                send('agent', {
-                  type: 'artifact',
-                  source: 'plain-stream',
-                  name: artifact.name,
-                  path: artifact.name,
-                  identifier: artifact.identifier,
-                  artifactType: artifact.artifactType,
-                });
-              }
-              send('agent', {
-                type: 'diagnostic',
-                name: 'plain_stream_artifacts_persisted',
-                source: 'daemon-run-finalize',
-                fileCount: persistedPlainArtifacts.length,
-                files: persistedPlainArtifacts.map((artifact) => artifact.name),
-              });
-            }
-          } catch (err) {
-            const message = err instanceof Error ? err.message : String(err);
-            const failureMessage = `Failed to persist plain-stream artifact(s): ${message}`;
-            console.warn(`[plain-stream] failed to persist stdout artifact(s): ${message}`);
-            send('agent', {
-              type: 'diagnostic',
-              name: 'plain_stream_artifacts_persist_failed',
-              source: 'daemon-run-finalize',
-              message,
-            });
-            send('error', createSseErrorPayload(
-              'AGENT_EXECUTION_FAILED',
-              failureMessage,
-            ));
-            return finishWithRetryDecision('failed', 1, null);
-          }
-        }
-      }
-      // Capture the pi session file path for conversational continuity.
-      // The session path is discovered by attachPiRpcSession when it
-      // processes agent_end; persist it under (conversationId, agentId) so
-      // another conversation in the same cwd cannot inherit this history.
-      if (acpSession && typeof acpSession.getLastSessionPath === 'function') {
-        const sessionPath = acpSession.getLastSessionPath();
-        if (status === 'succeeded' && def.streamFormat === 'pi-rpc') {
-          persistCapturedAgentSession(db, {
-            conversationId: run.conversationId,
-            agentId: def.id,
-            sessionId: sessionPath,
-            stablePromptHash: currentStableHash,
-            model: safeModel ?? null,
-            cwd: effectiveCwd,
-            lastMessageId: run.assistantMessageId ?? null,
-          });
-        }
-      }
-      // ACP session/load adapters (AMR/vela) report a durable upstream handle
-      // from the ACP session; persist it (under the resume-identity guard) so
-      // the next turn resumes via session/load. A missing handle clears the row
-      // (so a fresh session is opened next turn), mirroring the capture-style
-      // adapters.
-      if (
-        def.resumesSessionViaAcpLoad === true &&
-        status === 'succeeded' &&
-        acpSession &&
-        typeof acpSession.getDurableSessionId === 'function'
-      ) {
-        persistCapturedAgentSession(db, {
-          conversationId: run.conversationId,
-          agentId: def.id,
-          sessionId: acpSession.getDurableSessionId(),
-          stablePromptHash: currentStableHash,
-          model: safeModel ?? null,
-          cwd: effectiveCwd,
-          lastMessageId: run.assistantMessageId ?? null,
-        });
-      }
-      if (status === 'succeeded') {
-        try {
-          await snapshotAiHtmlVersionsBeforeSuccess();
-        } catch (err) {
-          const message = err instanceof Error ? err.message : String(err);
-          const details = err instanceof AiHtmlVersionSnapshotError
-            ? { failures: err.failures }
-            : undefined;
-          send('error', createSseErrorPayload(
-            'HTML_VERSION_SNAPSHOT_FAILED',
-            message,
-            {
-              retryable: false,
-              ...(details ? { details } : {}),
-            },
-          ));
-          design.runs.finish(run, 'failed', 1, signal);
-          return;
-        }
-        persistDeliveredAgentSessionState();
-      }
-      finishWithRetryDecision(status, code, signal);
-      } finally {
-        // Best-effort cleanup of the per-run agy log file on every close
-        // path — successful, failed, cancelled, or non-zero exit — so
-        // /tmp doesn't accumulate one file per Antigravity run. The log
-        // is read inside the empty-output guard above before this finally
-        // runs, so the read always happens before the unlink.
-        if (agentLogFilePath) {
-          fs.promises.unlink(agentLogFilePath).catch(() => {});
-        }
-        cleanupPromptFile();
-      }
-    });
-    if (writePromptToChildStdin && child.stdin) {
-      const promptInputFormat = def.promptInputFormat ?? 'text';
-      lifecycle.mark('model_call_start');
-      lifecycle.mark('stdin_write_start');
-      const markStdinWriteEnd = (err?: Error | null) => {
-        if (err) return;
-        lifecycle.mark('stdin_write_end');
-      };
-      if (promptInputFormat === 'stream-json') {
-        // Wrap the prompt as an Anthropic user message and write it as one
-        // JSONL line. Do NOT close stdin: claude-code keeps reading further
-        // messages until EOF, which is what lets the daemon stream more user
-        // messages into the same turn. The stdin is closed on a clean terminal
-        // turn (see applyClaudeStreamJsonRunBookkeeping) or when the child
-        // exits (run terminates, user cancels).
-        const userMessage = JSON.stringify({
-          type: 'user',
-          message: {
-            role: 'user',
-            content: [{ type: 'text', text: composed }],
-          },
-        });
-        try {
-          child.stdin.write(`${userMessage}\n`, 'utf8', markStdinWriteEnd);
-        } catch (err) {
-          // Swallow EPIPE here for the same reason as the listener above —
-          // a fast-exiting child has already routed its failure through
-          // stderr / exit handlers.
-          if (err && err.code !== 'EPIPE') throw err;
-        }
-        run.stdinOpen = true;
-      } else {
-        child.stdin.end(composed, 'utf8', markStdinWriteEnd);
-      }
-    }
-  };
+  const startChatRun = createStartChatRun({
+    ARTIFACTS_DIR,
+    DESIGN_SYSTEMS_DIR,
+    FORM_ANSWERED_GENERIC_OVERRIDE,
+    FORM_ANSWERED_SYSTEM_OVERRIDE,
+    FORM_ANSWERS_HEADER_RE,
+    OD_BIN,
+    OD_NODE_BIN,
+    PROJECTS_DIR,
+    PROJECT_ROOT,
+    RUNTIME_DATA_DIR,
+    SANDBOX_RUNTIME,
+    SKILLS_DIR,
+    activeChatAgentEventSinks,
+    activeChatRunHandles,
+    critiqueCfg,
+    critiqueRunRegistry,
+    critiqueWarnedAdapters,
+    runArtifactBaselines,
+    composeChatUserRequestForAgent,
+    createAgentRuntimeEnv,
+    createAgentRuntimeToolPrompt,
+    createAmrModelUnavailablePayload,
+    createSseErrorPayload,
+    emitProjectEvent,
+    filesystemEmptyAnswerFallbackText,
+    filesystemWriteFileNamesFromRunEvents,
+    hasGeneratedPluginArtifacts,
+    isPluginAuthoringRun,
+    persistRunEventToAssistantMessage,
+    refreshAndPersistToken,
+    renderRunContextPrompt,
+    resolveRunProjectKindForAnalytics,
+    rewriteKnownAgentStreamError,
+    scanRunEventsForRetrySideEffects,
+    telemetryPromptFromRunRequest,
+    composeDaemonSystemPrompt,
+    db,
+    design,
+    // live binding: server.ts reassigns daemonUrl after listen; the chat
+    // run must see the final URL, not the value at factory-creation time.
+    get daemonUrl() {
+      return daemonUrl;
+    },
+  });
 
   orbitService.setRunHandler(async ({
     trigger,
